% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R, R/Opal.R
\name{addGroupUser}
\alias{addGroupUser}
\alias{addGroupUser,missing-method}
\alias{addGroupUser,Opal-method}
\title{Add a user to a group}
\usage{
addGroupUser(object, group_id, user_id)

\S4method{addGroupUser}{missing}(object, group_id, user_id)

\S4method{addGroupUser}{Opal}(object, group_id, user_id)
}
\arguments{
\item{object}{An S4 object of class \link{Opal} that represents a connection to
the LMS.}

\item{group_id}{A character vector of length one specifying the group ID.}

\item{user_id}{A character vector of length one specifying the user ID.}
}
\value{
Status code of the HTTP request.

Status code \code{200} if the user was added successfully, or \code{NULL} when
OPAL rejects the request.
}
\description{
This method adds a user to a group on the Learning Management System (LMS).
If no LMS connection object is provided, it attempts to guess the connection
using default settings (e.g., environment variables). If the connection cannot
be established, an error is thrown.

This method adds a user to a learning group in LMS Opal by group id and user
id.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
addGroupUser("442662912", "196610")
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
addGroupUser("442662912", "196610")
\dontshow{\}) # examplesIf}
}
