% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{answerlist}
\alias{answerlist}
\title{Create markdown list for answer options}
\usage{
answerlist(vect, solutions = NULL)
}
\arguments{
\item{vect}{string or numeric vector of answer options for single/multiple
choice task}

\item{solutions}{numeric, optional; indexes of right answer options in \code{vect}}
}
\value{
markdown list
}
\description{
Create markdown list for answer options
}
