% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItemRef.R
\docType{class}
\name{AssessmentItemRef-class}
\alias{AssessmentItemRef-class}
\alias{AssessmentItemRef}
\title{Element asssessmentItemRef for xml testdescription}
\description{
Element assessmentItemRef for xml test description according to QTI 2.1
}
\section{Slots}{

\describe{
\item{\code{identifier}}{question identifier within a test}

\item{\code{href}}{path to xml file with assessementItem (task)}
}}

