% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentSection.R
\docType{class}
\name{AssessmentSection-class}
\alias{AssessmentSection-class}
\alias{AssessmentSection}
\title{Class AssessmentSection}
\description{
Class \code{AssessmentSection} is responsible for forming a section in test xml
specification according to QTI 2.1
}
\section{Slots}{

\describe{
\item{\code{identifier}}{section id}

\item{\code{title}}{title of the section}

\item{\code{visible}}{show or not to show for student}
}}

\examples{
essay <- new("Essay", prompt = "Test task", title = "Essay",
             identifier = "q1")
sc <- new("SingleChoice", prompt = "Test task", title = "SingleChoice",
          choices = c("A", "B", "C"), identifier = "q2")
exam_section <- new("AssessmentSection", identifier = "sec_id",
                   title = "section", assessment_item = list(essay, sc))
}
