% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R, R/Opal.R
\docType{methods}
\name{authLMS}
\alias{authLMS}
\alias{authLMS,Opal-method}
\title{Authenticate with LMS}
\usage{
authLMS(object, ...)

\S4method{authLMS}{Opal}(object, ...)
}
\arguments{
\item{object}{an instance of the S4 object \link{Opal}}

\item{...}{Additional arguments to be passed to the method, if applicable.}
}
\description{
A generic function to handle authentication with a Learning Management System
(LMS).

A generic function to handle authentication with LMS Opal.
}
