% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentSection.R, R/AssessmentItemRef.R
\docType{methods}
\name{buildAssessmentSection}
\alias{buildAssessmentSection}
\alias{buildAssessmentSection,AssessmentSection-method}
\alias{buildAssessmentSection,AssessmentSection}
\alias{buildAssessmentSection,AssessmentItemRef-method}
\alias{buildAssessementSection,AssessmentItemRef}
\title{Build tags for AssessmentSection in assessmentTest}
\usage{
buildAssessmentSection(object)

\S4method{buildAssessmentSection}{AssessmentSection}(object)

\S4method{buildAssessmentSection}{AssessmentItemRef}(object)
}
\arguments{
\item{object}{an instance of the S4 object (AssessmentSection, AssessmentItemRef)}
}
\description{
Generic function for tags that contains assessementSection in assessnetTest
}
