% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Choice.R
\docType{class}
\name{Choice-class}
\alias{Choice-class}
\alias{Choice}
\title{Class "Choice"}
\description{
Abstract class \code{Choice} is not meant to be instantiated directly; instead, it
serves as a base for derived classes \link[=SingleChoice-class]{SingleChoice} and
\link[=MultipleChoice-class]{MultipleChoice}.
}
\section{Slots}{

\describe{
\item{\code{choices}}{A character vector defining a set of answer options in the
question.}

\item{\code{choice_identifiers}}{A character vector, optional, containing a set of
identifiers for answers. By default, identifiers are generated
automatically according to the template "ChoiceD", where D is a letter
representing the alphabetical order of the answer in the list.}

\item{\code{shuffle}}{A boolean value indicating whether to randomize the order in
which the choices are initially presented to the candidate. Default is
\code{TRUE}.}

\item{\code{orientation}}{A character, determining whether to place answers in
vertical or horizontal mode. Possible values:
\itemize{
\item "vertical" - Default.
\item "horizontal"
}}
}}

