% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectFeedback.R
\name{correctFeedback}
\alias{correctFeedback}
\title{Create object \link{CorrectFeedback}}
\usage{
correctFeedback(content = list(), title = character(0), show = TRUE)
}
\arguments{
\item{content}{A character string or a list of character strings to form the
text of the question, which may include HTML tags.}

\item{title}{A character value, optional, representing the title of the
feedback window.}

\item{show}{A boolean value, optional, determining whether to show (\code{TRUE})
or hide (\code{FALSE}) the feedback. Default is \code{TRUE}.}
}
\value{
An object of class \link{CorrectFeedback}
}
\description{
Create object \link{CorrectFeedback}
}
\examples{
cfb <- correctFeedback(content = "Some comments", title = "Feedback")
}
