% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qti_test.R
\name{create_qti_test}
\alias{create_qti_test}
\title{Create XML file for exam test specification}
\usage{
create_qti_test(object, path = ".", verification = FALSE, zip_only
  = FALSE)
}
\arguments{
\item{object}{an instance of the \link{AssessmentTest} S4 object}

\item{path}{string, optional; a path to folder to store zip file with
possible file name; working directory by default}

\item{verification}{boolean, optional; checks the validity of the XML file.
Default is \code{FALSE}.}

\item{zip_only}{boolean, optional; if TRUE, returns only the zip file. If
FALSE, returns the zip, XML, and download files.}
}
\value{
xml document.
}
\description{
Create XML file for exam test specification
}
