% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentTest.R, R/AssessmentTestOpal.R
\docType{methods}
\name{createAssessmentTest}
\alias{createAssessmentTest}
\alias{createAssessmentTest,AssessmentTest-method}
\alias{createAssessmentTest,AssessmentTest}
\alias{createAssessmentTest,AssessmentTestOpal-method}
\alias{createAssessmentTest,AssessmentTestOpal}
\title{Create an element assessmentTest of a qti-xml document for test}
\usage{
createAssessmentTest(object, folder)

\S4method{createAssessmentTest}{AssessmentTest}(object, folder)

\S4method{createAssessmentTest}{AssessmentTestOpal}(object, folder)
}
\arguments{
\item{object}{an instance of the S4 object \link{AssessmentTest} or
\link{AssessmentTestOpal}}

\item{folder}{string, optional; a folder to store xml file; working directory
by default}
}
\description{
Generic function for creating assessmentTest element for XML document of
specification the test following the QTI schema v2.1
}
