% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentTest.R, R/AssessmentTestOpenOlat.R
\docType{methods}
\name{createConfigurationFile}
\alias{createConfigurationFile}
\alias{createConfigurationFile,AssessmentTest-method}
\alias{createConfigurationFile,AssessmentTest}
\alias{createConfigurationFile,AssessmentTestOpenOlat-method}
\alias{createConfigurationFile,AssessmentTestOpenOlat}
\title{Create a configuration file for QTI test}
\usage{
createConfigurationFile(object, output)

\S4method{createConfigurationFile}{AssessmentTest}(object, output)

\S4method{createConfigurationFile}{AssessmentTestOpenOlat}(object, output)
}
\arguments{
\item{object}{an instance of the S4 object \link{AssessmentTest}}

\item{output}{string, a folder to store an xml configuration files}
}
\description{
Generic function for creating additional configuration file for the archive with test
}
