% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/DirectedPair.R, R/Entry.R,
%   R/Essay.R, R/MultipleChoice.R, R/MultipleChoiceTable.R, R/OneInColTable.R,
%   R/OneInRowTable.R, R/Order.R, R/SingleChoice.R
\docType{methods}
\name{createItemBody}
\alias{createItemBody}
\alias{createItemBody,DirectedPair-method}
\alias{createItemBody,DirectedPair}
\alias{createItemBody,Entry-method}
\alias{createItemBody,Entry}
\alias{createItemBody,Essay-method}
\alias{createItemBody,Essay}
\alias{createItemBody,MultipleChoice-method}
\alias{createItemBody,MultipleChoice}
\alias{createItemBody,MultipleChoiceTable-method}
\alias{createItemBody,MultipleChoiceTable}
\alias{createItemBody,OneInColTable-method}
\alias{createItemBody,OneInColTable}
\alias{createItemBody,OneInRowTable-method}
\alias{createItemBody,OneInRowTable}
\alias{createItemBody,Order-method}
\alias{createItemBody,Order}
\alias{createItemBody,SingleChoice-method}
\alias{createItemBody,SingleChoice}
\title{Create an element itemBody of a qti-xml document}
\usage{
createItemBody(object)

\S4method{createItemBody}{DirectedPair}(object)

\S4method{createItemBody}{Entry}(object)

\S4method{createItemBody}{Essay}(object)

\S4method{createItemBody}{MultipleChoice}(object)

\S4method{createItemBody}{MultipleChoiceTable}(object)

\S4method{createItemBody}{OneInColTable}(object)

\S4method{createItemBody}{OneInRowTable}(object)

\S4method{createItemBody}{Order}(object)

\S4method{createItemBody}{SingleChoice}(object)
}
\arguments{
\item{object}{an instance of the S4 object (SingleChoice, MultipleChoice,
Entry, Order, OneInRowTable, OneInColTable, MultipleChoiceTable,
DirectedPair)}
}
\description{
Generic function for creating itemBody element for XML document of
specification the question following the QTI schema v2.1
}
