% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtiMetadata.R, R/AssessmentItem.R,
%   R/AssessmentTest.R
\docType{methods}
\name{createMetadata}
\alias{createMetadata}
\alias{createMetadata,QtiContributor-method}
\alias{createMetadata,QtiContributor}
\alias{createMetadata,AssessmentItem-method}
\alias{createMetadata,AssessmentItem}
\alias{createMetadata,AssessmentTest-method}
\alias{createMetadata,AssessmentTest}
\title{Create an element of metadata}
\usage{
createMetadata(object)

\S4method{createMetadata}{QtiContributor}(object)

\S4method{createMetadata}{AssessmentItem}(object)

\S4method{createMetadata}{AssessmentTest}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{QtiContributor}, \link{QtiMetadata}}
}
\description{
Create an element of metadata
}
