% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentTest.R,
%   R/Entry.R, R/Gap.R
\docType{methods}
\name{createOutcomeDeclaration}
\alias{createOutcomeDeclaration}
\alias{createOutcomeDeclaration,AssessmentItem-method}
\alias{createOutcomeDeclaration,AssessmentItem}
\alias{createOutcomeDeclaration,AssessmentTest-method}
\alias{createOutcomeDeclaration,AssessmentTest}
\alias{createOutcomeDeclaration,Entry-method}
\alias{createOutcomeDeclaration,Entry}
\alias{createOutcomeDeclaration,Gap-method}
\alias{createOutcomeDeclaration,TextGap}
\title{Create an element outcomeDeclaration of a qti-xml document}
\usage{
createOutcomeDeclaration(object)

\S4method{createOutcomeDeclaration}{AssessmentItem}(object)

\S4method{createOutcomeDeclaration}{AssessmentTest}(object)

\S4method{createOutcomeDeclaration}{Entry}(object)

\S4method{createOutcomeDeclaration}{Gap}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Ordering}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Generic function for creating outcomeDeclaration element for XML document of
specification the question following the QTI schema v2.1
}
