% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentTest.R
\docType{methods}
\name{createQtiTest-methods}
\alias{createQtiTest-methods}
\alias{createQtiTest}
\alias{createQtiTest,AssessmentTest-method}
\alias{createQtiTest,AssessmentTest}
\alias{createQtiTest,character-method}
\alias{createQtiTest,character}
\title{Create XML file for exam test specification}
\usage{
createQtiTest(object, dir = NULL, verification = FALSE, zip_only =
  FALSE)

\S4method{createQtiTest}{AssessmentTest}(object, dir = getwd(), verification = FALSE, zip_only = FALSE)

\S4method{createQtiTest}{character}(object, dir = getwd())
}
\arguments{
\item{object}{an instance of the \link{AssessmentTest} or \link{AssessmentTestOpal} S4
object}

\item{dir}{string, optional; a folder to store xml file; working directory by
default}

\item{verification}{boolean, optional; to check validity of xml file, default
\code{FALSE}}

\item{zip_only}{boolean, optional; returns only zip file in case of TRUE or
zip, xml and downloads files in case of FALSE value; FALSE by default}
}
\value{
xml document.
}
\description{
Create XML file for exam test specification
}
\examples{
\dontrun{
essay <- new("Essay", prompt = "Test task", title = "Essay",
             identifier = "q1")
sc <- new("SingleChoice", prompt = "Test task", title = "SingleChoice",
          choices = c("A", "B", "C"), identifier = "q2")
exam_section <- new("AssessmentSection", identifier = "sec_id",
                   title = "section", assessment_item = list(essay, sc))
exam <- new("AssessmentTestOpal", identifier = "id_test",
           title = "some title", section = list(exam_section))
createQtiTest(exam, "exam_folder", "TRUE")
}
}
