% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/MatchTable.R, R/Entry.R,
%   R/Essay.R, R/InlineChoice.R, R/MultipleChoice.R, R/MultipleChoiceTable.R,
%   R/NumericGap.R, R/Order.R, R/SingleChoice.R, R/TextGap.R
\docType{methods}
\name{createResponseDeclaration}
\alias{createResponseDeclaration}
\alias{createResponseDeclaration,AssessmentItem-method}
\alias{createResponseDeclaration,AssessmentItem}
\alias{createResponseDeclaration,MatchTable-method}
\alias{createResponseDeclaration,MatchTable}
\alias{createResponseDeclaration,Entry-method}
\alias{createResponseDeclaration,Entry}
\alias{createResponseDeclaration,Essay-method}
\alias{createResponseDeclaration,Essay}
\alias{createResponseDeclaration,InlineChoice-method}
\alias{createResponseDeclaration,InlineChoice}
\alias{createResponseDeclaration,MultipleChoice-method}
\alias{createResponseDeclaration,MultipleChoice}
\alias{createResponseDeclaration,MultipleChoiceTable-method}
\alias{createResponseDeclaration,MultipleChoiceTable}
\alias{createResponseDeclaration,NumericGap-method}
\alias{createResponseDeclaration,NumericGap}
\alias{createResponseDeclaration,Order-method}
\alias{createResponseDeclaration,Order}
\alias{createResponseDeclaration,SingleChoice-method}
\alias{createResponseDeclaration,SingleChoice}
\alias{createResponseDeclaration,TextGap-method}
\alias{createResponseDeclaration,TextGap}
\title{Create an element responseDeclaration of a qti-xml document}
\usage{
createResponseDeclaration(object)

\S4method{createResponseDeclaration}{AssessmentItem}(object)

\S4method{createResponseDeclaration}{MatchTable}(object)

\S4method{createResponseDeclaration}{Entry}(object)

\S4method{createResponseDeclaration}{Essay}(object)

\S4method{createResponseDeclaration}{InlineChoice}(object)

\S4method{createResponseDeclaration}{MultipleChoice}(object)

\S4method{createResponseDeclaration}{MultipleChoiceTable}(object)

\S4method{createResponseDeclaration}{NumericGap}(object)

\S4method{createResponseDeclaration}{Order}(object)

\S4method{createResponseDeclaration}{SingleChoice}(object)

\S4method{createResponseDeclaration}{TextGap}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Generic function for creating responseDeclaration element for XML document of
specification the question following the QTI schema v2.1
}
