% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_processing.R
\name{createTemplateDeclaration}
\alias{createTemplateDeclaration}
\alias{createTemplateDeclaration,MaximaVariables-method}
\alias{createTemplateDeclaration,AssessmentItem-method}
\title{Create QTI template declarations}
\usage{
createTemplateDeclaration(object)

\S4method{createTemplateDeclaration}{MaximaVariables}(object)

\S4method{createTemplateDeclaration}{AssessmentItem}(object)
}
\arguments{
\item{object}{A \link{MaximaVariables} or \link{AssessmentItem} object.}
}
\value{
An \code{htmltools} tag list, or \code{NULL} for an item without templates.
}
\description{
Create QTI template declarations
}
