% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dropdown}
\alias{dropdown}
\title{Create YAML string for InlineChoice object (dropdown list)}
\usage{
dropdown(
  choices,
  solution_index = 1,
  points = 1,
  shuffle = TRUE,
  response_identifier = NULL,
  choices_identifiers = NULL
)
}
\arguments{
\item{choices}{numeric or character vector; contains values of possible
answers}

\item{solution_index}{integer, optional; the number of right answer in the
\code{choices} vector, default \code{1}}

\item{points}{numeric, optional; the number of points for this gap; default
\code{1}}

\item{shuffle}{boolean, optional; is responsible to randomize the order in
which the choices are initially presented to the candidate; default \code{TRUE}}

\item{response_identifier}{string; an identifier for the answer; by default
it is generated automatically}

\item{choices_identifiers}{character vector, optional; a set of identifiers
for answers; by default identifiers are generated automatically}
}
\value{
string; map yaml
}
\description{
Create YAML string for InlineChoice object (dropdown list)
}
