% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_integration.R
\name{exams_task}
\alias{exams_task}
\title{Convert an exams Rmd task to a QTI 2.1 XML file}
\usage{
exams_task(path, title = NULL)
}
\arguments{
\item{path}{Character string. Path to an Rmd file written in the \code{exams}
format.}

\item{title}{Character string or \code{NULL}. Optional title passed to
\code{\link[exams:exams2qti21]{exams::exams2qti21()}} via the \code{ititle} argument.}
}
\value{
Character string. Path to the QTI 2.1 XML file generated by
\code{\link[exams:exams2qti21]{exams::exams2qti21()}} in a temporary directory.
}
\description{
Helper function for using tasks written in the \code{exams} Rmd format inside
\code{rqti} sections. The function converts the given Rmd file with
\code{\link[exams:exams2qti21]{exams::exams2qti21()}} and returns the path to the generated XML file.
}
\details{
This function is mainly useful when \code{exams} tasks need to be combined with
native \code{rqti} tasks in one section, for example:
\code{section(exams_task(path_to_file, title = "Exams task"))}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
path <- system.file("exams", "example.Rmd", package = "rqti")
section(exams_task(path, title = "Exams task"))
\dontshow{\}) # examplesIf}
}
