% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_result_zip}
\alias{extract_result_zip}
\title{Create data frame with test results from Opal zip file}
\usage{
extract_result_zip(file, details = "answers")
}
\arguments{
\item{file}{A string with a path to the zip file with results}

\item{details}{A string with two possible values:
\itemize{
\item "answers" - default, returns data frame with answers
\item "options" - returns data frame with options
}}
}
\value{
data frame.
}
\description{
The function \code{extract_result_zip()} creates data frames from opal zip file
}
