% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{fetch_resource_url}
\alias{fetch_resource_url}
\title{Create a URL using the resource's display name in LMS OPAL}
\usage{
fetch_resource_url(
  display_name,
  endpoint = "https://bildungsportal.sachsen.de/opal/",
  api_user = NULL,
  api_password = NULL
)
}
\arguments{
\item{display_name}{character; target display_name}

\item{endpoint}{endpoint of LMS Opal; default is
"https://bildungsportal.sachsen.de/opal/"}

\item{api_user}{username on OPAL}

\item{api_password}{password on OPAL}
}
\description{
Create a URL using the resource's display name in LMS OPAL
}
