% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gap_numeric}
\alias{gap_numeric}
\title{Create YAML string for NumericGap object}
\usage{
gap_numeric(
  solution,
  tolerance = 0,
  tolerance_type = "absolute",
  points = 1,
  response_identifier = NULL,
  include_lower_bound = TRUE,
  include_upper_bound = TRUE,
  expected_length = size_gap(solution),
  placeholder = NULL
)
}
\arguments{
\item{solution}{A numeric value; contains the correct answer for this numeric
entry.}

\item{tolerance}{A numeric value, optional; specifies the upper and lower
boundaries of the tolerance range for the candidate's answer. Default is 0.}

\item{tolerance_type}{A character string, optional; specifies tolerance mode;
possible values:"exact", "absolute" (by default), "relative".}

\item{points}{A numeric value, optional; the number of points for this gap.
Default is 1.}

\item{response_identifier}{A character string, optional; an identifier for
the answer.}

\item{include_lower_bound}{A boolean, optional; specifies whether the
lower bound is included in the tolerance range.}

\item{include_upper_bound}{A boolean, optional; specifies whether the
upper bound is included in the tolerance range.}

\item{expected_length}{An integer value, optional; sets the size of the text
input field in the content delivery engine.}

\item{placeholder}{A character string, optional; places helpful text in the
text input field in the content delivery engine.}
}
\value{
A character string mapped as YAML.
}
\description{
Create YAML string for NumericGap object
}
\examples{
gap_numeric(5.0, tolerance = 10, tolerance_type = "relative")

}
\seealso{
\code{\link[=gap_text]{gap_text()}}, \code{\link[=dropdown]{dropdown()}}, \code{\link[=mdlist]{mdlist()}}
}
