% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_processing.R
\name{german_grading}
\alias{german_grading}
\title{German grading scale}
\usage{
german_grading()
}
\value{
A named numeric vector with grades as names and minimum
score proportions as values.
}
\description{
A helper function that returns a named numeric vector representing
a common German grading scheme. The names correspond to grades and
the values define the minimum proportion of points required to
achieve the respective grade.
}
\examples{
german_grading()


}
