% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{get_result_attr_answers}
\alias{get_result_attr_answers}
\title{Create data frame with test results by answers}
\usage{
get_result_attr_answers(file)
}
\arguments{
\item{file}{A string with a path of the xml test result file}
}
\value{
data frame.
}
\description{
The function \code{get_result_attr_answers()} creates data frames with the
following data about the test results:
\itemize{
\item 'file_name' - name of the xml file with test results (to identify
candidate)
\item 'datestamp' - date and time of test
\item 'question_id' - question item identifier
\item 'duration' - time in sec. what candidate spent on this item
\item 'candidate_score' - points that were given to candidate after evaluation
\item 'max_scored' - max possible score for this question item
\item 'type' - the type of question
}
}
