% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{get_result_attr_options}
\alias{get_result_attr_options}
\title{Create data frame with test results by options of answers}
\usage{
get_result_attr_options(file)
}
\arguments{
\item{file}{A string with a path of the xml test result file}
}
\value{
data frame.
}
\description{
The function \code{get_result_attr_options()} creates data frames with the
following data about the individual answers:
\itemize{
\item 'file_name' - name of the xml file with test results (to identify
candidate)
\item 'datestamp' - date and time of test
\item 'question_id' - question item identifier
\item 'base_types' - type of answer (identifier, string or float)
\item 'cardinalities' - defines whether this question is single, multiple or
ordered -value
\item 'qti_type' - specifies the type of the task
\item 'id_answer_options' - identifier of each response variable
\item 'correct_responses' - values that considered as right responses for
question
\item 'cand_responses' - values that were given by candidate
}
}
