% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{get_result_attributes}
\alias{get_result_attributes}
\title{Create data frame with test rusults}
\usage{
get_result_attributes(file)
}
\arguments{
\item{file}{A string with a path of the xml test result file}
}
\value{
data frame.
}
\description{
The function 'get_result_attributes()'creates data frames with the following
data about the test results: 'identifier' - question item identifier,
'duration' - time in sec. what candidate spent on this item, 'score' - points
that were given to candidate after evaluation, "max_scored' - max poissible
score for this item, 'type' - the type of question, 'candidate_answer' -
identifiers of choosen options or specific answer (or NA in case there was no
response from candidate), 'correct_responses' - identifiers of options or
specific answer that consider as a right response for this item
}
