% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentSection.R, R/AssessmentItemRef.R
\docType{methods}
\name{getAssessmentItems}
\alias{getAssessmentItems}
\alias{getAssessmentItems,AssessmentSection-method}
\alias{getAssessmentItems,AssessmentSection}
\alias{getAssessmentItems,AssessmentItemRef-method}
\alias{getAssessmentItems,AssessmentItemRef}
\title{Get list of AssessmentItems for AssessmentSection}
\usage{
getAssessmentItems(object)

\S4method{getAssessmentItems}{AssessmentSection}(object)

\S4method{getAssessmentItems}{AssessmentItemRef}(object)
}
\arguments{
\item{object}{an instance of the S4 object (AssessmentSection)}
}
\description{
Generic function for
}
