% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentSection.R,
%   R/character.R
\docType{methods}
\name{getCalculator-methods}
\alias{getCalculator-methods}
\alias{getCalculator}
\alias{getCalculator,AssessmentItem-method}
\alias{getCalculator,AssessmentItem}
\alias{getCalculator,AssessmentSection-method}
\alias{getCalculator,AssessmentSection}
\alias{getCalculator,character-method}
\alias{getCalculator,character}
\title{Get value of the slot 'calculator'}
\usage{
getCalculator(object)

\S4method{getCalculator}{AssessmentItem}(object)

\S4method{getCalculator}{AssessmentSection}(object)

\S4method{getCalculator}{character}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Ordering}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Get value of the slot 'calculator'
}
