% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opal.R
\name{getCourseElements,Opal-method}
\alias{getCourseElements,Opal-method}
\title{Retrieve Data About Course Elements from LMS Opal}
\usage{
\S4method{getCourseElements}{Opal}(object, course_id)
}
\arguments{
\item{object}{An S4 object of class \link{Opal} that represents a connection to the LMS.}

\item{course_id}{A length one character vector with course id.}
}
\value{
A dataframe with the data of the elements of the course (fields: nodeId,
shortTitle, shortName, longTitle) on LMS Opal.
}
\description{
This function retrieves and returns data about the elements of a specified course
from the LMS Opal system. The data includes information such as the node ID,
short title, short name, and long title of each element.
}
