% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{getCourseElements}
\alias{getCourseElements}
\title{Get elements of the course by courseId from LMS}
\usage{
getCourseElements(object, course_id)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the LMS.}

\item{course_id}{A length one character vector with course id.}
}
\value{
A dataframe with the elements of the course.
}
\description{
Get elements of the course by courseId from LMS
}
