% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R, R/Opal.R
\name{getCourseGroups}
\alias{getCourseGroups}
\alias{getCourseGroups,missing-method}
\alias{getCourseGroups,Opal-method}
\title{Get groups from a course}
\usage{
getCourseGroups(object, course_id)

\S4method{getCourseGroups}{missing}(object, course_id)

\S4method{getCourseGroups}{Opal}(object, course_id)
}
\arguments{
\item{object}{An S4 object of class \link{Opal} that represents a connection to
the LMS.}

\item{course_id}{A character vector of length one specifying the course
resource ID. Note that this is not the course ID shown in the URL, but a
longer identifier available via the "Show more information" option within
the course.}
}
\value{
A data frame with all accessible attributes of the course groups.

A data frame with course groups and their attributes.
}
\description{
This method retrieves groups from a course on the Learning Management System (LMS)
by its course id. It returns a data frame with all accessible attributes of the
groups.
If no LMS connection object is provided, it attempts to guess the connection using
default settings (e.g., environment variables). If the connection cannot be
established, an error is thrown.

This method retrieves groups from a course on LMS Opal by its course id. The
groups are returned as a data frame with all accessible attributes, such as
group id, name, description, participant limits, and group settings. If no
Opal connection object is provided, it attempts to guess the connection using
default settings (e.g., environment variables). If the connection cannot be
established, an error is thrown.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
groups <- getCourseGroups("89068111333293")
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
groups <- getCourseGroups("89068111333293")
\dontshow{\}) # examplesIf}
}
