% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opal.R
\name{getCourseResult,Opal-method}
\alias{getCourseResult,Opal-method}
\title{Get zip with course results by resource id and node id}
\usage{
\S4method{getCourseResult}{Opal}(
  object,
  resource_id,
  node_id,
  path_outcome = ".",
  rename = TRUE
)
}
\arguments{
\item{object}{An S4 object of class \link{Opal} that represents a connection to the LMS.}

\item{resource_id}{A length one character vector with resource id.}

\item{node_id}{A length one character vector with node id (test).}

\item{path_outcome}{A length one character vector with path, where the zip should be
stored. Default is working directory.}

\item{rename}{A boolean value; optional; Set \code{TRUE} value to take the short
name of the course element for naming zip (results_shortName.zip). \code{FALSE}
combines in zip name course id and node id. Default is \code{TRUE}.}
}
\value{
It downloads a zip and return a character string with path.
}
\description{
Get zip with course results by resource id and node id
}
