% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{getCourseResult}
\alias{getCourseResult}
\title{Get zip with course results by resource id and node id}
\usage{
getCourseResult(object, resource_id, node_id, path_outcome = ".", ...)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the LMS.}

\item{resource_id}{A length one character vector with resource id.}

\item{node_id}{A length one character vector with node id (test).}

\item{path_outcome}{A length one character vector with path, where the zip should be
stored. Default is working directory.}

\item{...}{Additional arguments to be passed to the method, if applicable.}
}
\value{
It downloads a zip and return a character string with path.
}
\description{
Get zip with course results by resource id and node id
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
zip_file <- getCourseResult("89068111333293", "1617337826161777006")
\dontshow{\}) # examplesIf}
}
