% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{getCourseUsers}
\alias{getCourseUsers}
\alias{getCourseUsers,missing-method}
\title{Get users from a course}
\usage{
getCourseUsers(object, course_id)

\S4method{getCourseUsers}{missing}(object, course_id)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the LMS.}

\item{course_id}{A length one character vector with course id.}
}
\value{
A data frame with course participants. Each row represents a participant,
including an additional column indicating the corresponding group.
}
\description{
This method retrieves users from a course on the Learning Management System (LMS)
by its course id. The users are collected from all groups within the course and
returned as a single data frame. If no LMS connection object is provided, it attempts
to guess the connection using default settings (e.g., environment variables).
If the connection cannot be established, an error is thrown.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
users <- getCourseUsers("89068111333293")
\dontshow{\}) # examplesIf}
}
