% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R, R/Opal.R
\name{getGroupUsers}
\alias{getGroupUsers}
\alias{getGroupUsers,missing-method}
\alias{getGroupUsers,Opal-method}
\title{Get users from a group}
\usage{
getGroupUsers(object, group_id)

\S4method{getGroupUsers}{missing}(object, group_id)

\S4method{getGroupUsers}{Opal}(object, group_id)
}
\arguments{
\item{object}{An S4 object of class \link{Opal} that represents a connection to the LMS.}

\item{group_id}{A character vector specifying the group id.}
}
\value{
A data frame with all accessible attributes of the group members.

A data frame with group users and their attributes. Each row represents
a user and may include the following columns:
\describe{
\item{key}{Unique user id.}
\item{login}{User login name.}
\item{firstName}{User first name.}
\item{lastName}{User last name.}
\item{email}{User email address.}
\item{status}{User status (e.g., \code{STATUS_ACTIVE}).}
}
}
\description{
This method retrieves users from a group on the Learning Management System (LMS)
by its group id. It returns a data frame with all accessible attributes of the
group members. If no LMS connection object is provided, it attempts to guess
the connection using default settings (e.g., environment variables). If the
connection cannot be established, an error is thrown.

This method retrieves users from a group on LMS Opal by its group id.
It returns a data frame with user attributes obtained from the LMS,
including user id, login, first name, last name, email, and additional
properties (e.g., status).
If no Opal connection object is provided, it attempts to guess the
connection using default settings (e.g., environment variables).
If the connection cannot be established, an error is thrown.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
groups <- getGroupUsers("89068111333293")
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
users <- getGroupUsers("89068111333293")
\dontshow{\}) # examplesIf}
}
