% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{getLMSResources,missing-method}
\alias{getLMSResources,missing-method}
\title{Get records of all current user's resources on LMS}
\usage{
\S4method{getLMSResources}{missing}(object)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the
LMS.}
}
\value{
A dataframe with attributes of user's resources.
}
\description{
This function retrieves data about all resources associated with the current user on the Learning Management System (LMS).
If no LMS connection object is provided, it attempts to guess the connection using default settings (e.g., environment variables).
If the connection cannot be established, an error is thrown.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- getLMSResources()
\dontshow{\}) # examplesIf}
}
