% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opal.R
\name{getLMSResources,Opal-method}
\alias{getLMSResources,Opal-method}
\title{Get records of all current user's resources on LMS Opal}
\usage{
\S4method{getLMSResources}{Opal}(object)
}
\arguments{
\item{object}{An S4 object of class \link{Opal} that represents a connection to the LMS.}
}
\value{
A dataframe with attributes of user's resources.
}
\description{
Get records of all current user's resources on LMS Opal
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- getLMSResources()
\dontshow{\}) # examplesIf}
}
