% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{getLMSResources}
\alias{getLMSResources}
\title{Get records of all current user's resources on LMS}
\usage{
getLMSResources(object)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the
LMS.}
}
\value{
A dataframe with attributes of user's resources.
}
\description{
Get records of all current user's resources on LMS
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- getLMSResources()
\dontshow{\}) # examplesIf}
}
