% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opal.R
\name{getLMSResourcesByName,Opal-method}
\alias{getLMSResourcesByName,Opal-method}
\title{Get selected records of the current user's resources on LMS Opal by display
name.}
\usage{
\S4method{getLMSResourcesByName}{Opal}(object, display_name, rtype = NULL)
}
\arguments{
\item{object}{An S4 object of class \link{Opal} that represents a connection to
the LMS.}

\item{display_name}{A string value withe the name of resource.}

\item{rtype}{A string value with the type of resource. Possible values:
"FileResource.TEST", "FileResource.QUESTION", or "FileResource.SURVEY".}
}
\value{
A dataframe with attributes of user's resources.
}
\description{
Get selected records of the current user's resources on LMS Opal by display
name.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- getLMSResourcesByName()
\dontshow{\}) # examplesIf}
}
