% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{getLMSResourcesByName}
\alias{getLMSResourcesByName}
\title{Get select records about user resources by name.}
\usage{
getLMSResourcesByName(object, display_name, rtype = NULL)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the
LMS.}

\item{display_name}{A string value withe the name of resource.}

\item{rtype}{A string value with the type of resource. Possible values:
"FileResource.TEST", "FileResource.QUESTION", or "FileResource.SURVEY".}
}
\value{
A dataframe with attributes of user's resources.
}
\description{
Get select records about user resources by name.
}
