% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentSection.R,
%   R/character.R
\docType{methods}
\name{getObject-methods}
\alias{getObject-methods}
\alias{getObject}
\alias{getObject,AssessmentItem-method}
\alias{getObject,AssessmentItem}
\alias{getObject,AssessmentSection-method}
\alias{getObject,AssessmentSection}
\alias{getObject,character-method}
\alias{getObject,character}
\title{Get object}
\usage{
getObject(object)

\S4method{getObject}{AssessmentItem}(object)

\S4method{getObject}{AssessmentSection}(object)

\S4method{getObject}{character}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Get object
}
