% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R
\docType{methods}
\name{getPoints}
\alias{getPoints}
\title{Get points from AssessmentItem object}
\usage{
getPoints(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Get points from AssessmentItem object
}
