% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gap.R, R/InlineChoice.R, R/NumericGap.R,
%   R/TemplateVariableRef.R, R/TextGap.R, R/character.R
\docType{methods}
\name{getResponse}
\alias{getResponse}
\alias{getResponse,InlineChoice-method}
\alias{getResponse,InlineChoice}
\alias{getResponse,NumericGap-method}
\alias{getResponse,NumericGap}
\alias{getResponse,TemplateVariableRef-method}
\alias{getResponse,TemplateVariableRef}
\alias{getResponse,TextGap-method}
\alias{getResponse,TextGap}
\alias{getResponse,character-method}
\alias{getResponse,character}
\title{Get and process a piece of question content}
\usage{
getResponse(object)

\S4method{getResponse}{InlineChoice}(object)

\S4method{getResponse}{NumericGap}(object)

\S4method{getResponse}{TemplateVariableRef}(object)

\S4method{getResponse}{TextGap}(object)

\S4method{getResponse}{character}(object)
}
\arguments{
\item{object}{an instance of the S4 object (NumericGap, TextGap,
InlineChoice, character)}
}
\description{
Generic function to get and process a different types of question content
(text with instances of gaps or dropdown lists) for XML document of
specification the question following the QTI schema v2.1
}
