% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Img.R
\docType{class}
\name{Img-class}
\alias{Img-class}
\alias{Img}
\title{Class Img}
\description{
Class \code{Img} is responsible to keep attributes and process images as a part of
excericise content
}
\section{Slots}{

\describe{
\item{\code{src}}{string; uri of picture}

\item{\code{alt}}{string, optional; an alternate text for an image, if the image
cannot be displayed}

\item{\code{width}}{numeric, optional; specifies the width of an image, in pixels}

\item{\code{height}}{numeric, optional; specifies the height of an image, in pixels}
}}

