% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InlineChoice.R
\docType{class}
\name{InlineChoice-class}
\alias{InlineChoice-class}
\alias{InlineChoice}
\title{Class InlineChoice}
\description{
Abstract class \code{InlineChoice} is responsible for creating instances of
dropdown lists as answer options in question Entry type assessment task
according to QTI 2.1.
}
\section{Slots}{

\describe{
\item{\code{response_identifier}}{string; an identifier for thy answer; by default it
is generated automatically}

\item{\code{points}}{numeric, optional; the number of points for this gap;
default \code{1}}

\item{\code{placeholder}}{string, optional; is responsible to place some helpful text
in text input field in content delivery engine}

\item{\code{expected_length}}{numeric, optional; is responsible to set a size of text
input field in content delivery engine}

\item{\code{choices}}{string vector; contains the answers which are shown in the drop
down list}

\item{\code{solution_index}}{numeric, optional; the index of the right answer in options
vector; default \code{1}}

\item{\code{choices_identifiers}}{string vector, optional; a set of identifiers for
answers; by default identifiers are generated automatically}

\item{\code{shuffle}}{boolean, optional; is responsible to randomize the order in
which the choices are initially presented to the candidate; default \code{TRUE}}
}}

\examples{
ng <- new("InlineChoice",
          choices =  c("answer1", "answer2", "answer3"),
          response_identifier = "dropdown_gap_example")
}
