% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_functions.R
\name{knit_qti_html}
\alias{knit_qti_html}
\title{Generate a HTML representation of a RMD file}
\usage{
knit_qti_html(input, ...)
}
\arguments{
\item{input}{(the path to the input Rmd document)}

\item{...}{other arguments that are currently ignored}
}
\description{
Choosing a specific Rmd file the function generates the qti.xml for it through the functions of the package exams, after the XML is copied into the QTI Engine which transforms the qti.xml into HTML format. Finally, the HTML is displayed and the user can have a preview of the exercise or exam.
}
\examples{
# Customize knit function in the Rmd file using the following YAML setting after the word knit:
qtiViewer::knit_qti_html
}
