% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalFeedback.R
\name{modalFeedback}
\alias{modalFeedback}
\title{Create object \link{ModalFeedback}}
\usage{
modalFeedback(content = list(), title = character(0), show = TRUE)
}
\arguments{
\item{content}{A list of character content to form the text of the modal
feedback, which can include HTML tags.}

\item{title}{A character value, optional, representing the title of the modal
feedback window.}

\item{show}{A boolean value, optional, determining whether to show (\code{TRUE}) or
hide (\code{FALSE}) the modal feedback. Default is \code{TRUE}.}
}
\value{
An object of class \link{ModalFeedback}
}
\description{
Create object \link{ModalFeedback}
}
\examples{
fb <- modalFeedback(content = list("Model answer"), title = "Feedback")
}
