% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqti.R
\name{normalize_html_content}
\alias{normalize_html_content}
\title{Normalize HTML content}
\usage{
normalize_html_content(content)
}
\description{
Recursively converts objects created by \code{htmltools} to character strings.
Other content is returned unchanged.
}
