% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentSection.R,
%   R/AssessmentTest.R, R/character.R
\docType{methods}
\name{prepareQTIJSFiles-methods}
\alias{prepareQTIJSFiles-methods}
\alias{prepareQTIJSFiles}
\alias{prepareQTIJSFiles,AssessmentItem-method}
\alias{prepareQTIJSFiles,AssessmentItem}
\alias{prepareQTIJSFiles,AssessmentSection-method}
\alias{prepareQTIJSFiles,AssessmentSection}
\alias{prepareQTIJSFiles,AssessmentTest-method}
\alias{prepareQTIJSFiles,AssessmentTest}
\alias{prepareQTIJSFiles,character-method}
\title{Prepare files to render them with QTIJS}
\usage{
prepareQTIJSFiles(object, dir = NULL)

\S4method{prepareQTIJSFiles}{AssessmentItem}(object, dir = "")

\S4method{prepareQTIJSFiles}{AssessmentSection}(object, dir = NULL)

\S4method{prepareQTIJSFiles}{AssessmentTest}(object, dir = NULL)

\S4method{prepareQTIJSFiles}{character}(object, dir = NULL)
}
\arguments{
\item{object}{an instance of \link{AssessmentItem}, \link{AssessmentTest},
\link{AssessmentTestOpal}, \link{AssessmentSection}, or string path to xml, rmd or md
files}

\item{dir}{QTIJS path}
}
\description{
Prepare files to render them with QTIJS
}
