% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_qti.R
\name{print.qti_validation_result}
\alias{print.qti_validation_result}
\title{Print a QTI validation result}
\usage{
\method{print}{qti_validation_result}(x, ...)
}
\arguments{
\item{x}{A \code{qti_validation_result} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a QTI validation result
}
