% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_qti.R
\name{print.qti_validation_results_list}
\alias{print.qti_validation_results_list}
\title{Print validation results list}
\usage{
\method{print}{qti_validation_results_list}(x, ...)
}
\description{
Print validation results list
}
\keyword{internal}
