% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{publishCourse,missing-method}
\alias{publishCourse,missing-method}
\title{Publish a course on LMS}
\usage{
\S4method{publishCourse}{missing}(object, course_id)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the LMS.}

\item{course_id}{A character string with course id in the LMS.}
}
\value{
Status code of the HTTP request.
}
\description{
This method publishes the course by its course id on Learning Management System (LMS).
If no LMS connection object is provided, it attempts to guess the connection using default settings (e.g., environment variables).
If the connection cannot be established, an error is thrown.
}
