% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opal.R
\name{publishCourse,Opal-method}
\alias{publishCourse,Opal-method}
\title{Publish a course on LMS Opal}
\usage{
\S4method{publishCourse}{Opal}(object, course_id)
}
\arguments{
\item{object}{An S4 object of class \link{Opal} that represents a connection to the LMS.}

\item{course_id}{A character string with course id in the LMS.}
}
\value{
Status code of the HTTP request.
}
\description{
Publish a course on LMS Opal
}
