% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{publishCourse}
\alias{publishCourse}
\title{Publish a course on LMS}
\usage{
publishCourse(object, course_id)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the LMS.}

\item{course_id}{A character string with course id in the LMS.}
}
\value{
Status code of the HTTP request.
}
\description{
Publish a course on LMS
}
