% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtiMetadata.R
\name{qti_metadata}
\alias{qti_metadata}
\title{Constructor function for class QtiMetadata}
\usage{
qti_metadata(
  contributor = list(),
  description = "",
  rights = Sys.getenv("QTI_RIGHTS"),
  version = "0.0.9"
)
}
\arguments{
\item{contributor}{A list of objects \link{QtiContributor}-type that holds
metadata information about the authors.}

\item{description}{A character string providing a textual description of the
content of this learning object.}

\item{rights}{A character string describing the intellectual property rights
and conditions of use for this learning object. By default it takes value
from environment variable 'QTI_RIGHTS'.}

\item{version}{A character string representing the edition/version of this
learning object.}
}
\description{
Creates object of \link{QtiMetadata}-class
}
\examples{
creator= qti_metadata(qti_contributor("Max Mustermann"),
                      description = "Task description",
                      rights = "This file is Copyright (C) 2024 Max
                      Mustermann, all rights reserved.",
                      version = "1.0")
}
