% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_functions.R
\name{qtijs_pkg_path}
\alias{qtijs_pkg_path}
\title{Shortcut for the qtijs path of the rqti package local installation.}
\usage{
qtijs_pkg_path()
}
\description{
Shortcut for the qtijs path of the rqti package local installation.
}
\examples{
  qtijs_pkg_path()
}
