% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtiMetadata.R
\docType{class}
\name{QtiMetadata-class}
\alias{QtiMetadata-class}
\alias{QtiMetadata}
\title{Class QtiMetadata}
\description{
This class stores metadata information such as author name, description,
rights, version, format, creation date, and modified date for QTI-compliant
tasks or tests.
}
\section{Slots}{

\describe{
\item{\code{author}}{A character string representing the name of the author.}

\item{\code{description}}{A character string providing a textual description of the
content of this learning object.}

\item{\code{rights}}{A character string describing the intellectual property rights
and conditions of use for this learning object.}

\item{\code{version}}{A character string representing the edition/version of this
learning object.}

\item{\code{format}}{A character string representing the QTI (Question and Test
Interoperability) information model version. Default is 'IMS QTI 2.1'.}

\item{\code{created_date}}{A character string representing the creation date of the
learning object. Default is the current system date.}

\item{\code{modified_date}}{A character string representing the date of the last
modification of the learning object. Default is the current system date.}
}}

