% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectBuilder.R
\name{rmd2opal}
\alias{rmd2opal}
\title{Upload task on LMS OPAL}
\usage{
rmd2opal(file)
}
\arguments{
\item{file}{a file with markdown description of the question.}
}
\description{
Direct uploading a task from Rmd file on OPAL.
}
