% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectBuilder.R
\name{rmd2qti}
\alias{rmd2qti}
\title{Create qti-XML task file from Rmd (md) description}
\usage{
rmd2qti(file, dir = NULL, verification = FALSE)
}
\arguments{
\item{file}{a file with markdown description of question.}

\item{dir}{string, optional; a folder to store xml file; working directory by
default}

\item{verification}{boolean, optional; to check validity of xml file, default
\code{FALSE}}
}
\value{
xml document
}
\description{
Create XML file for question specification from Rmd (md) description
according to qti 2.1 infromation model
}
