% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_builder.R
\name{rmd2xml}
\alias{rmd2xml}
\title{Create qti-XML task file from Rmd (md) description}
\usage{
rmd2xml(file, path = getwd(), verification = FALSE)
}
\arguments{
\item{file}{A string of path to file with markdown description of question.}

\item{path}{A string, optional; a folder to store xml file. Default is
working directory.}

\item{verification}{A boolean value, optional; enable validation of the xml
file. Default is \code{FALSE}.}
}
\value{
The path string to the xml file.
}
\description{
Create XML file for question specification from Rmd (md) description
according to qti 2.1 infromation model
}
\examples{
\dontrun{
# creates folder with xml (side effect)
rmd2xml("task.Rmd", "target_folder", TRUE)
}
}
