% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_builder.R
\name{section}
\alias{section}
\title{Create a section as a part of test content}
\usage{
section(file, num_variants = 1, id = NULL, nested = TRUE)
}
\arguments{
\item{file}{vector of Rmd, md, or xml files}

\item{num_variants}{number of variants from Rmd files}

\item{id}{identifier of section}

\item{nested}{the type of the test structure}
}
\description{
Create a section as a part of test content
}
