% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateVariableRef.R
\docType{class}
\name{TemplateVariableRef-class}
\alias{TemplateVariableRef-class}
\alias{TemplateVariableRef}
\title{Reference to a template variable in item content}
\description{
Use this object in an assessment item's \code{content} list to display a template
variable using QTI \code{printedVariable}. It does not define an answer field or
contribute points. When the complete item is exported, the referenced name
must be declared in its \code{template} list. Both named outputs and intermediate
result identifiers can be referenced. Rmd parsing is not yet supported.
}
\section{Slots}{

\describe{
\item{\code{identifier}}{A single valid QTI identifier of the template variable.}
}}

\examples{
variables <- new("MaximaVariables", code = "string(['a = 1]);",
                 variables = c(a = "integer"))
reference <- new("TemplateVariableRef", identifier = "a")
item <- new("Entry", template = list(variables),
            content = list("<p>Enter ", reference, ": ",
                           numericGap(1), "</p>"))
createItemBody(item)
}
