% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextGap.R
\docType{class}
\name{TextGap-class}
\alias{TextGap-class}
\alias{TextGap}
\title{Class TextGap}
\description{
Abstract class \code{TextGap} is responsible for creating instances of input
fields with text type of answer in question Entry type assessment task
according to QTI 2.1.
}
\section{Slots}{

\describe{
\item{\code{response_identifier}}{string; an identifier for thy answer; by default it
is generated automatically}

\item{\code{score}}{numeric, optional; the number of points for this gap;
default \code{1}}

\item{\code{placeholder}}{string, optional; is responsible to place some helpful text
in text input field in content delivery engine}

\item{\code{expected_length}}{numeric, optional; is responsible to set a size of text
input field in content delivery engine}

\item{\code{response}}{string; contains right answer for this text gap entry}

\item{\code{alternatives}}{string vector, optional; contains a vector of values that are
also considered correct answers}

\item{\code{case_sentensive}}{logical; determines whether the evaluation of the
correct answer is case sensitive; default \code{TRUE}}
}}

\examples{
tg <- new("TextGap",
          response = "answer",
          alternatives = c("answerr", "aanswer"),
          placeholder = "do not put special characters" )
}
