% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{textgapopal}
\alias{textgapopal}
\title{Create YAML string for TextGapOpal object}
\usage{
textgapopal(
  response,
  alternatives = NULL,
  score = NULL,
  expected_length = NULL,
  placeholder = NULL,
  case_sensitive = NULL,
  response_identifier = NULL,
  value_precision = NULL
)
}
\arguments{
\item{response}{string; contains right answer for this text gap entry}

\item{alternatives}{string vector, optional; contains a vector of values that
are also considered correct answers}

\item{score}{numeric, optional; the number of points for this gap; default 1}

\item{expected_length}{numeric, optional; is responsible to set a size of
text input field in content delivery engine}

\item{placeholder}{string, optional; is responsible to place some helpful
text in text input field in content delivery engine}

\item{case_sensitive}{logical, optional; determines whether the evaluation of
the correct answer is case sensitive}

\item{response_identifier}{string; an identifier for thy answer; by default
it is generated automatically}

\item{value_precision}{numeric, optional; defines how many characters will be
taken into account to tolerate spelling mistake in evaluation of candidate
answer}
}
\value{
string; map yaml
}
\description{
Create YAML string for TextGapOpal object
}
