% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{updateCourseElementResource,missing-method}
\alias{updateCourseElementResource,missing-method}
\title{Update the referenced learning resource of a course element in the LMS}
\usage{
\S4method{updateCourseElementResource}{missing}(object, course_id, ...)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the LMS.}

\item{course_id}{A character string with course id in the LMS.}

\item{...}{Additional arguments to be passed to the method, if applicable.}
}
\value{
Response of the HTTP request.
}
\description{
This method updates the learning resource by its course id on Learning Management System (LMS).
If no LMS connection object is provided, it attempts to guess the connection using default settings (e.g., environment variables).
If the connection cannot be established, an error is thrown.
}
