% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{upload_opal_test}
\alias{upload_opal_test}
\title{Upload test on OPAL}
\usage{
upload_opal_test(file, display_name = NULL)
}
\arguments{
\item{file}{required; a length one character vector}

\item{display_name}{optional; a length one character vector to entitle file in
OPAL; file name without extension by default}
}
\value{
status code
}
\description{
Function \code{upload_opal_test()} takes full prepared zip archive of QTI-test and
uploads it to the OPAL. before calling \code{upload_opal_test()} authentication
procedure has been performed. See \link{auth_opal}
}
