% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrongFeedback.R
\docType{class}
\name{WrongFeedback-class}
\alias{WrongFeedback-class}
\alias{WrongFeedback}
\title{Class "WrongFeedback"}
\description{
Class \code{WrongFeedback} is responsible for delivering feedback messages
to the candidate in case of an incorrect answer on the entire exercise.
}
\section{Slots}{

\describe{
\item{\code{outcome_identifier}}{A character representing the unique identifier of
the outcome declaration variable that relates to feedback. Default is
"FEEDBACKMODAL".}

\item{\code{show}}{A boolean value, optional, determining whether to show (\code{TRUE}) or
hide (\code{FALSE}) the modal feedback. Default is \code{TRUE}.}

\item{\code{title}}{A character value, optional, representing the title of the modal
feedback window.}

\item{\code{content}}{A list of character content to form the text of the modal
feedback, which can include HTML tags.}

\item{\code{identifier}}{A character value representing the identifier of the modal
feedback item. Default is "incorrect".}
}}

\examples{
wfb <- new("WrongFeedback",
          title = "Wrong answer",
          content = list("<b>Some demonstration</b>"))
}
