% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrongFeedback.R
\name{wrongFeedback}
\alias{wrongFeedback}
\title{Create object \link{WrongFeedback}}
\usage{
wrongFeedback(content = list(), title = character(0), show = TRUE)
}
\arguments{
\item{content}{A list of character content to form the text of the feedback,
which can include HTML tags.}

\item{title}{A character value, optional, representing the title of the
feedback window.}

\item{show}{A boolean value, optional, determining whether to show (\code{TRUE})
or hide (\code{FALSE}) the feedback. Default is \code{TRUE}.}
}
\value{
An object of class \link{WrongFeedback}
}
\description{
Create object \link{WrongFeedback}
}
\examples{
wfb <- wrongFeedback(content = list("Some comments"), title = "Feedback")
}
