% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectBuilder.R
\name{xml2opal}
\alias{xml2opal}
\title{Upload xml-file with task on LMS OPAL}
\usage{
xml2opal(file)
}
\arguments{
\item{file}{a xml-file with QTI task specification}
}
\description{
Direct uploading a task from xml file on OPAL.
}
