% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R
\docType{class}
\name{AssessmentItem-class}
\alias{AssessmentItem-class}
\title{Root element assessmentItem for xml task description}
\description{
Root element assessmentItem for xml task description accroding to QTI 2.1
}
\section{Slots}{

\describe{
\item{\code{text}}{an object Text with}

\item{\code{points}}{Number of points for the whole task}

\item{\code{title}}{Title of the file}

\item{\code{identifier}}{file id}

\item{\code{qti_version}}{qti information model version}
}}

