% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R
\docType{class}
\name{AssessmentItem-class}
\alias{AssessmentItem-class}
\alias{AssessmentItem}
\title{Class AssessmentItem}
\description{
Abstract class \code{AssessmentItem} is responsible for creating a root element
'assessmentItem' in xml task description according to QTI 2.1. This class is
never to be generated, only derived classes meaningful.
}
\section{Slots}{

\describe{
\item{\code{identifier}}{string; identifier of assessment task}

\item{\code{title}}{string; a title of xml file}

\item{\code{prompt}}{string, optional; simple question text, consisting of one
paragraph, can supplement or replace content}

\item{\code{qti_version}}{string; QTI information model version; default 'v2p1}

\item{\code{feedback}}{list of \linkS4class{ModalFeedback}, \link{CorrectFeedback}, or
\linkS4class{WrongFeedback}; feedback messages for candidate}

\item{\code{files}}{string, optional; the names of the attachment files that should
be available in the test}

\item{\code{calculator}}{string, optional; determines to show to candidate calculator
during the test;
possible values:
\itemize{
\item "simple-calculator"
\item "scientific-calculator"
\item "" - by default
}}
}}

\section{Warning}{
 This class is not useful in itself, but some classes derive
from it.
}

