% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentSection.R
\docType{class}
\name{AssessmentSection-class}
\alias{AssessmentSection-class}
\title{Root element asssessmentSection for xml task description}
\description{
Root element assessmentSection for xml task description according to QTI 2.1
}
\section{Slots}{

\describe{
\item{\code{identifier}}{section id}

\item{\code{title}}{title of the section}

\item{\code{visible}}{show or not to show for student}
}}

