% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentTest.R
\docType{class}
\name{AssessmentTest-class}
\alias{AssessmentTest-class}
\title{Root element assessmentTest for xml task description}
\description{
Root element assessmentTest for xml task description according to QTI 2.1
}
\section{Slots}{

\describe{
\item{\code{identifier}}{test file id}

\item{\code{title}}{Title of the file}

\item{\code{qti_version}}{qti information model version}
}}

