% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{auth_opal}
\alias{auth_opal}
\title{Authentification in OPAL API}
\usage{
auth_opal(api_user = NULL, api_password = NULL, endpoint = NULL)
}
\arguments{
\item{api_user}{username on OPAL}

\item{api_password}{password on OPAL}

\item{endpoint}{endpoint of LMS Opal; by default it is got from environment
variable \code{QTI_API_ENDPOINT}. To set a global environment variable, you need
to call \code{Sys.setenv(QTI_API_ENDPOINT='xxxxxxxxxxxxxxx')} or you can put
these command into .Renviron.}
}
\value{
user id
}
\description{
Function \code{auth_opal()} performs the necessary authentication steps in OPAL
API. If the authentication is successful, the function sets the token value
in the system environment and returns the user's identity key in OPAL. The
token value is required to access the OPAL API system.
}
\section{Authentication}{
 To use OPAL API, you need to provide your
OPAL-username and password. This package uses system credential store
'keyring' to store user's name and password. After the first successful
access to the OPAL API, there is no need to specify the username and
password again, they will be extracted from the system credential store
}

