% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{auth_opal}
\alias{auth_opal}
\title{Authentification in OPAL API}
\usage{
auth_opal()
}
\value{
user id
}
\description{
Function \code{auth_opal()} performs the necessary authentication steps in OPAL
API. If the authentication is successful, the function sets the cookie value
in the system environment and returns the status code. The cookie value is
required to access the OPAL API system."
}
\section{Authentication}{
 To use OPAL API, you need to provide your OPAL-
username and password. This function will look for API_USER and
API_PASSWORD environment variables. To set a global environment variables,
you need to use the following commands:
\code{Sys.setenv(API_USER ='xxxxxxxxxxxxxxx')}
\code{Sys.setenv(API_PASSWORD =  'xxxxxxxxxxxxxxx')}
Another way to assign environment variables in case of regular using is to
create a file named .env in the root directory of your project. The .env
file should contain the environment variables you want to set in the
following format:
\code{API_USER=xxxxxxxxxxxxxxx}
\code{API_PASSWORD=xxxxxxxxxxxxxxx}
}

