% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{auth_opal}
\alias{auth_opal}
\title{Authentification in OPAL API}
\usage{
auth_opal(api_user = NULL, api_password = NULL, endpoint = NULL, cached = TRUE)
}
\arguments{
\item{api_user}{username on OPAL}

\item{api_password}{password on OPAL}

\item{endpoint}{endpoint of LMS Opal; default is set from environment variable
\code{QTI_API_ENDPOINT}. To set a global environment variable, you need to call \verb{ }Sys.setenv(QTI_API_ENDPOINT='xxxxxxxxxxxxxxx')` or you can put these
commands into .Renviron.}

\item{cached}{boolean; if \code{TRUE} (default) it keeps password in environment
variable}
}
\value{
user id
}
\description{
Function \code{auth_opal()} performs the necessary authentication steps in OPAL
API. If the authentication is successful, the function sets the cookie value
in the system environment and returns the user's identity key in OPAL. The
cookie value is required to access the OPAL API system."
}
\section{Authentication}{
 To use OPAL API, you need to provide your
OPAL-username and password. This function can get \code{QTI_API_USER} and
\code{QTI_API_PASSWORD} from environment variables. To set a global environment
variable, you need to call \code{Sys.setenv(QTI_API_USER ='xxxxxxxxxxxxxxx')} and
\code{Sys.setenv(QTI_API_PASSWORD ='xxxxxxxxxxxxxxx')}or you can put these commands
into .Rprofile.
}

