% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\docType{methods}
\name{authLMS}
\alias{authLMS}
\alias{authLMS,LMS-method}
\title{Authenticate with LMS}
\usage{
authLMS(object, ...)

\S4method{authLMS}{LMS}(object, ...)
}
\arguments{
\item{object}{an instance of the S4 object \link{Opal}, \link{LMS}}

\item{...}{Additional arguments to be passed to the method, if applicable.}
}
\description{
A generic function to handle authentication with a Learning Management System
(LMS).

The method to handle authentication with LMS.
}
