% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentSection.R, R/AssessmentItemRef.R,
%   R/character.R
\docType{methods}
\name{buildAssessmentSection}
\alias{buildAssessmentSection}
\alias{buildAssessmentSection,AssessmentSection-method}
\alias{buildAssessmentSection,AssessmentSection}
\alias{buildAssessmentSection,AssessmentItem-method}
\alias{buildAssessementSection,AssessmentItem}
\alias{buildAssessmentSection,AssessmentItemRef-method}
\alias{buildAssessementSection,AssessmentItemRef}
\alias{buildAssessmentSection,character-method}
\alias{buildAssessementSection,character}
\title{Build tags for AssessmentSection in assessmentTest}
\usage{
buildAssessmentSection(object, folder = NULL)

\S4method{buildAssessmentSection}{AssessmentSection}(object, folder = NULL)

\S4method{buildAssessmentSection}{AssessmentItem}(object, folder = NULL)

\S4method{buildAssessmentSection}{AssessmentItemRef}(object)

\S4method{buildAssessmentSection}{character}(object, folder = NULL)
}
\arguments{
\item{object}{an instance of the S4 object (\link{AssessmentSection},
\link{AssessmentItemRef} and all types of \link{AssessmentItem})}

\item{folder}{string; a folder to store xml file}
}
\description{
Generic function for tags that contains assessementSection in assessnetTest
}
