% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentSection.R,
%   R/character.R
\docType{methods}
\name{buildAssessmentSection}
\alias{buildAssessmentSection}
\alias{buildAssessmentSection,AssessmentItem-method}
\alias{buildAssessementSection,AssessmentItem}
\alias{buildAssessmentSection,AssessmentSection-method}
\alias{buildAssessmentSection,AssessmentSection}
\alias{buildAssessmentSection,character-method}
\alias{buildAssessementSection,character}
\title{Build tags for AssessmentSection in assessmentTest}
\usage{
buildAssessmentSection(object, folder = NULL, verify = FALSE)

\S4method{buildAssessmentSection}{AssessmentItem}(object, folder)

\S4method{buildAssessmentSection}{AssessmentSection}(object, folder = NULL, verify = FALSE)

\S4method{buildAssessmentSection}{character}(object, folder = NULL, verify = FALSE)
}
\arguments{
\item{object}{an instance of the S4 object (\link{AssessmentSection} and all types
of \link{AssessmentItem})}

\item{folder}{string; a folder to store xml file}

\item{verify}{boolean, optional; check validity of xml file, default \code{FALSE}}
}
\description{
Generic function for tags that contains assessementSection in assessnetTest
}
