% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Choice.R
\docType{class}
\name{Choice-class}
\alias{Choice-class}
\alias{Choice}
\title{Class "Choice"}
\description{
Abstract class \code{Choice} is never to be generated, only derived classes
\linkS4class{SingleChoice} and \linkS4class{MultipleChoice} are meaningful.
}
\section{Slots}{

\describe{
\item{\code{choices}}{string vector; defines a set of answers options in question}

\item{\code{shuffle}}{boolean, optional; is responsible to randomize the order in
which the choices are initially presented to the candidate; default \code{TRUE}}

\item{\code{choice_identifiers}}{string vector, optional; a set of identifiers for
answers; by default identifiers are generated automatically}

\item{\code{orientation}}{string, optional; is responsible to place answers in
vertical or horizontal mode; possible values:
\itemize{
\item "vertical" - default
\item "horizontal"
}}
}}

\section{Warning}{
 This class is not useful in itself, but some classes derive
from it.
}

