% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectFeedback.R
\docType{class}
\name{CorrectFeedback-class}
\alias{CorrectFeedback-class}
\alias{CorrectFeedback}
\title{Class CorrectFeedback}
\description{
Abstract class \code{CorrectFeedback} is responsible to deliver feedback messages
to candidate in case of correct answer on the whole exercise
}
\section{Slots}{

\describe{
\item{\code{outcome_identifier}}{string; identifier if outcome declaration variable
that relates to feedback}

\item{\code{show}}{boolean, optional; is responsible to show (\code{TRUE} value) or hide
(\code{FALSE}); default \code{TRUE}}

\item{\code{identifier}}{string; identifier of model feedback item; default 'modal
feedback'}

\item{\code{title}}{string, optional; title of model feedback item}

\item{\code{content}}{list; list of strings with formatting and images (\linkS4class{Img})
to form the text of the question}
}}

