% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qti.R
\name{create_qti_task}
\alias{create_qti_task}
\title{Create XML file for question specification}
\usage{
create_qti_task(object,
                dir = NULL,
                verification = FALSE)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable}, \link{MultipleChoiceTable},
\link{DirectedPair}).}

\item{dir}{string, optional; a folder to store xml file; working directory by
default}

\item{verification}{boolean, optional; to check validity of xml file, default
\code{FALSE}}
}
\value{
xml document.
}
\description{
Create XML file for question specification
}
\examples{
\dontrun{
essay <- new("Essay", prompt = "Test task", title = "Essay")
create_qti_task(essay, "result", "TRUE")
}
}
