% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qti.R
\name{create_qti_task}
\alias{create_qti_task}
\title{Create XML file for question specification}
\usage{
create_qti_task(object, path = getwd())
}
\arguments{
\item{object}{an instance of the S4 object (SingleChoice, MultipleChoice,
Entry, Order, OneInRowTable, OneInColTable, MultipleChoiceTable,
DirectedPair).}

\item{path}{a folder to store xml file}
}
\value{
xml document.
}
\description{
Create XML file for question specification
}
