% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qti_test.R
\name{create_qti_test}
\alias{create_qti_test}
\title{Create XML file for exam test specification}
\usage{
create_qti_test(object,
                dir = NULL,
                verification = FALSE)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable}, \link{MultipleChoiceTable},
\link{DirectedPair}).}

\item{dir}{string, optional; a folder to store xml file; working directory by
default}

\item{verification}{boolean, optional; to check validity of xml file, default
\code{FALSE}}
}
\value{
xml document.
}
\description{
Create XML file for exam test specification
}
\examples{
\dontrun{
essay <- new("Essay", prompt = "Test task", title = "Essay",
             identifier = "q1")
sc <- new("SingleChoice", prompt = "Test task", title = "SingleChoice",
          choices = c("A", "B", "C"), identifier = "q2")
exam <- new("AssessmentTest", )
create_qti_test(essay, "result", "TRUE")
}
}
