% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qti_test.R
\name{create_qti_test}
\alias{create_qti_test}
\title{Create XML file for exam test specification}
\usage{
create_qti_test(object, path = getwd(), verification = FALSE)
}
\arguments{
\item{object}{an instance of the \link{AssessmentTest} S4 object}

\item{path}{string, optional; a path to folder to store zip file with possible
file name; working directory by default}

\item{verification}{boolean, optional; to check validity of xml file, default
\code{FALSE}}
}
\value{
xml document.
}
\description{
Create XML file for exam test specification
}
\examples{
\dontrun{
essay <- new("Essay", prompt = "Test task", title = "Essay",
             identifier = "q1")
sc <- new("SingleChoice", prompt = "Test task", title = "SingleChoice",
          choices = c("A", "B", "C"), identifier = "q2")
exam_section <- new("AssessmentSection", identifier = "sec_id",
                   title = "section", assessment_item = list(essay, sc))
exam <- new("AssessmentTestOpal", identifier = "id_test",
           title = "some title", section = list(exam_section))
create_qti_test(exam, "exam_folder/exam.zip", "TRUE")
}
}
