% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_builder.R
\name{create_question_object}
\alias{create_question_object}
\title{Create rqti S4 \link{AssessmentItem} Object from Rmd}
\usage{
create_question_object(file)
}
\arguments{
\item{file}{A string representing the path to an Rmd file.}
}
\value{
One of the rqti S4 AssessmentItem objects: \link{SingleChoice},
\link{MultipleChoice}, \link{Essay}, \link{Entry}, \link{Ordering}, \link{OneInRowTable},
\link{OneInColTable}, \link{MultipleChoiceTable}, or \link{DirectedPair}.
}
\description{
Generates an rqti S4 AssessmentItem object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Ordering}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}) from an Rmd file.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
create_question_object("file.Rmd")
\dontshow{\}) # examplesIf}
}
