% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectBuilder.R
\name{create_question_object}
\alias{create_question_object}
\title{Create S4 object according to content of markdown file}
\usage{
create_question_object(file)
}
\arguments{
\item{file}{a file with markdown description of question.}
}
\value{
an instance of the S4 object (SingleChoice, MultipleChoice,
Entry, Order, OneInRowTable, OneInColTable, MultipleChoiceTable,
DirectedPair).
}
\description{
Create S4 object according to content of markdown file
}
