% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentTest.R
\docType{methods}
\name{createAssessmentTest}
\alias{createAssessmentTest}
\title{Create an element assessmentTest of a qti-xml document for test}
\usage{
createAssessmentTest(object)
}
\arguments{
\item{object}{an instance of the S4 object (AssessmentTest)}
}
\description{
Generic function for creating assessmentTest element for XML document of
specification the test following the QTI schema v2.1
}
