% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/Entry.R, R/InlineChoice.R,
%   R/MultipleChoice.R, R/NumericGap.R, R/TextGap.R
\docType{methods}
\name{createOutcomeDeclaration}
\alias{createOutcomeDeclaration}
\alias{createOutcomeDeclaration,AssessmentItem-method}
\alias{createOutcomeDeclaration,AssessmentItem}
\alias{createOutcomeDeclaration,Entry-method}
\alias{createOutcomeDeclaration,Entry}
\alias{createOutcomeDeclaration,InlineChoice-method}
\alias{createOutcomeDeclaration,InlineChoice}
\alias{createOutcomeDeclaration,MultipleChoice-method}
\alias{createOutcomeDeclaration,MultipleChoice}
\alias{createOutcomeDeclaration,NumericGap-method}
\alias{createOutcomeDeclaration,NumericGap}
\alias{createOutcomeDeclaration,TextGap-method}
\alias{createOutcomeDeclaration,TextGap}
\title{Create an element outcomeDeclaration of a qti-xml document}
\usage{
createOutcomeDeclaration(object)

\S4method{createOutcomeDeclaration}{AssessmentItem}(object)

\S4method{createOutcomeDeclaration}{Entry}(object)

\S4method{createOutcomeDeclaration}{InlineChoice}(object)

\S4method{createOutcomeDeclaration}{MultipleChoice}(object)

\S4method{createOutcomeDeclaration}{NumericGap}(object)

\S4method{createOutcomeDeclaration}{TextGap}(object)
}
\arguments{
\item{object}{an instance of the S4 object (AssessmentItem, Entry,
InlineChoice, MultipleChoice, NumericGap, TextGap)}
}
\description{
Generic function for creating outcomeDeclaration element for XML document of
specification the question following the QTI schema v2.1
}
