% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R
\docType{methods}
\name{createQtiTask-methods}
\alias{createQtiTask-methods}
\alias{createQtiTask}
\alias{createQtiTask,AssessmentItem-method}
\alias{createQtiTask,AssessmentItem}
\title{Create XML file for question specification}
\usage{
createQtiTask(object, dir = NULL, verification = FALSE)

\S4method{createQtiTask}{AssessmentItem}(object, dir = NULL, verification = FALSE)
}
\arguments{
\item{object}{An instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}).}

\item{dir}{A character value, optional; a folder to store xml file; working
directory is used by default.}

\item{verification}{A boolean value, optional; to check validity of xml file.
Default is \code{FALSE}.}
}
\value{
xml document.
}
\description{
Create XML file for question specification
}
\examples{
\dontrun{
essay <- new("Essay", prompt = "Test task", title = "Essay")
createQtiTask(essay, "result", TRUE)
}
}
