% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/Choice.R, R/Entry.R,
%   R/Gap.R, R/InlineChoice.R, R/NumericGap.R, R/Order.R, R/TextGap.R,
%   R/TextGapOpal.R
\docType{methods}
\name{createResponseProcessing}
\alias{createResponseProcessing}
\alias{createResponseProcessing,Choice-method}
\alias{createResponseProcessing,Choice}
\alias{createResponseProcessing,Entry-method}
\alias{createResponseProcessing,Entry}
\alias{createResponseProcessing,Gap-method}
\alias{createResponseProcessing,Gap}
\alias{createResponseProcessing,InlineChoice-method}
\alias{createResponseProcessing,InlineChoice}
\alias{createResponseProcessing,NumericGap-method}
\alias{createResponseProcessing,NumericGap}
\alias{createResponseProcessing,Order-method}
\alias{createResponseProcessing,Order}
\alias{createResponseProcessing,TextGap-method}
\alias{createResponseProcessing,TextGap}
\alias{createResponseProcessing,TextGapOpal-method}
\alias{createResponseProcessing,TextGapOpal}
\title{Create an element responseProcessing of a qti-xml document}
\usage{
createResponseProcessing(object)

\S4method{createResponseProcessing}{Choice}(object)

\S4method{createResponseProcessing}{Entry}(object)

\S4method{createResponseProcessing}{Gap}(object)

\S4method{createResponseProcessing}{InlineChoice}(object)

\S4method{createResponseProcessing}{NumericGap}(object)

\S4method{createResponseProcessing}{Order}(object)

\S4method{createResponseProcessing}{TextGap}(object)

\S4method{createResponseProcessing}{TextGapOpal}(object)
}
\arguments{
\item{object}{an instance of the S4 object (Choice, Entry, Gap, InlineChoice,
NumericGap, Order)}
}
\description{
Generic function for creating responseProcessing element for XML document of
specification the question following the QTI schema v2.1
}
