% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentTest.R
\docType{methods}
\name{createTestPart}
\alias{createTestPart}
\alias{createTestPart,AssessmentTest-method}
\alias{createTestPart,AssessmentTest}
\title{Create an element testPart of a qti-xml document for test}
\usage{
createTestPart(object)

\S4method{createTestPart}{AssessmentTest}(object)
}
\arguments{
\item{object}{an instance of the S4 object (AssessmentTest)}
}
\description{
Generic function for creating testPart element for XML document of
specification the test following the QTI schema v2.1
}
