% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entry.R, R/Gap.R, R/InlineChoice.R,
%   R/character.R
\docType{methods}
\name{createText}
\alias{createText}
\alias{createText,Gap-method}
\alias{createText,Gap}
\alias{createText,InlineChoice-method}
\alias{createText,InlineChoice}
\alias{createText,character-method}
\alias{createText,character}
\title{Compose a set of html elements to display question in qti-xml document}
\usage{
createText(object)

\S4method{createText}{Gap}(object)

\S4method{createText}{InlineChoice}(object)

\S4method{createText}{character}(object)
}
\arguments{
\item{object}{an instance of the S4 object (Gap, InlineChoice, character)}
}
\description{
Generic function for creating a set of html elements to display question for
XML document of specification the question following the QTI schema v2.1
}
