% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentTest.R, R/AssessmentTestOpal.R
\docType{methods}
\name{createZip}
\alias{createZip}
\alias{createZip,AssessmentTest-method}
\alias{createZip,AssessmentTest}
\alias{createZip,AssessmentTestOpal-method}
\alias{createZip,AssessmentTestOpal}
\title{Create an Zip archive of QTI test}
\usage{
createZip(object, folder)

\S4method{createZip}{AssessmentTest}(object, folder)

\S4method{createZip}{AssessmentTestOpal}(object, folder)
}
\arguments{
\item{object}{an instance of the S4 object \link{AssessmentTest} or
\link{AssessmentTestOpal}}

\item{folder}{string, optional; a folder to store xml file; working directory
by default}
}
\description{
Generic function for creating zip archive with set of XML documents of
specification the test following the QTI schema v2.1
}
