% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entry.R
\docType{class}
\name{Entry-class}
\alias{Entry-class}
\alias{Entry}
\title{Class "Entry"}
\description{
Abstract class \code{Entry} is responsible for creating assessment task according
to QTI 2.1., where question includes one or more instances of text input
fields (with numeric or text answers) or dropdown lists.
}
\section{Slots}{

\describe{
\item{\code{content}}{list; a list of content to form text of question, includes text
(characters) and instances of Gap-objects: \linkS4class{TextGap}, \link{NumericGap},
\link{InlineChoice}}

\item{\code{identifier}}{string; identifier of assessment task}

\item{\code{title}}{string; a title of xml file}

\item{\code{prompt}}{string, optional; simple question text, consisting of one
paragraph, can supplement or replace content}

\item{\code{qti_version}}{string; QTI information model version; default 'v2p1}

\item{\code{feedback}}{vector of \linkS4class{ModalFeedback}; feedback messages for
candidate}

\item{\code{points}}{numeric, optional; the number of points for the entire task;
default 1}
}}

\note{
use \code{\link[=create_qti_task]{create_qti_task()}} to create an xml file of a task specification
}
\examples{
entry_gaps <- new("Entry", content = list("<p>In mathematics, the common
logarithm is the logarithm with base", new("NumericGap",
                                           response_identifier = "numeric_1",
                                           response = 10,
                                           placeholder = "it is a number"),
". It is also known as the decimal", new("TextGap",
                                         response_identifier = "text_1",
                                         response = "logarithm",
                                         placeholder = "it is a text"),
 ".</p>"),
                   title = "entry with number and text in answers",
                   identifier = "entry_example")
entry_dropdown <- new("Entry", content = list("<p>In mathematics, the common
logarithm is the logarithm with base", new("InlineChoice",
                                           response_identifier = "numeric_1",
                                           options = c("10", "7", "11")),
". It is also known as the decimal", new("InlineChoice",
                                         response_identifier = "text_1",
                                         options = c("logarithm", "limit")),
 ".</p>"),
                   title = "entry with dropdown lists for answers",
                   identifier = "entry_example")
}
