% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_result_zip}
\alias{extract_result_zip}
\title{Create data frame with test results from Opal zip file}
\usage{
extract_result_zip(file, details = "answers")
}
\arguments{
\item{file}{A string with a path to the zip file with results}

\item{details}{A string with two possible values:
\itemize{
\item "answers" - default, returns data frame with answers; see data frame
structure at \code{\link[=get_result_attr_answers]{get_result_attr_answers()}}
\item "options" - returns data frame with options; see data frame structure at
[get_result_attr_options()
}}
}
\value{
data frame.
}
\description{
The function \code{extract_result_zip()} creates data frames from opal zip file.
}
\note{
final data frame contains additional column 'titles' with values of
attribute 'title' of assessment item
of the tasks
}
