% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_results}
\alias{extract_results}
\title{Create data frame with test results}
\usage{
extract_results(file, level = "exercises")
}
\arguments{
\item{file}{A string with a path of the xml test result file}

\item{level}{A string with two possible options: exercises and items}
}
\value{
data frame.
}
\description{
The function \code{extract_results()} takes Opal zip archive "Export results" or
xml file and creates two kinds of data frames (according
to parameter 'level')
\enumerate{
\item with optioin level = "excercises" data set consists of columns:
}
\itemize{
\item 'file_name' - name of the xml file with test results (to identify
candidate)
\item 'datestamp' - date and time of test
\item 'question_id' - question item identifier
\item 'duration' - time in sec. what candidate spent on this item
\item 'candidate_score' - points that were given to candidate after evaluation
\item 'max_scored' - max possible score for this question item
\item 'question_type' - the type of question
\item 'is_answer_given' - TRUE if candidate gave the answer on question,
otherwise FALSE
\item 'titles' - the values of attribute 'title' of assessment items
}
\enumerate{
\item with optioin level = "items" data set consists of columns:
}
\itemize{
\item 'file_name' - name of the xml file with test results (to identify
candidate)
\item 'datestamp' - date and time of test
\item 'question_id' - question item identifier
\item 'base_types' - type of answer (identifier, string or float)
\item 'cardinalities' - defines whether this question is single, multiple or
ordered -value
\item 'qti_type' - specifies the type of the task
\item 'id_answer_options' - identifier of each response variable
\item 'correct_responses' - values that considered as right responses for
question
\item 'cand_responses' - values that were given by candidate
\item 'titles' - the values of attribute 'title' of assessment items
}
}
