% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gap.R
\docType{class}
\name{Gap-class}
\alias{Gap-class}
\alias{Gap}
\title{Class "Gap"}
\description{
Abstract class \code{Gap} is never to be generated, only derived classes
\linkS4class{TextGap}, \link{NumericGap} and \link{InlineChoice} are meaningful.
}
\section{Warning}{
 This class is not useful in itself, but some classes derive
from it.
}

