% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gap_numeric}
\alias{gap_numeric}
\title{Create YAML string for NumericGap object}
\usage{
gap_numeric(
  solution,
  score = NULL,
  expected_length = NULL,
  placeholder = NULL,
  tolerance = NULL,
  tolerance_type = NULL,
  include_lower_bound = NULL,
  include_upper_bound = NULL,
  response_identifier = NULL
)
}
\arguments{
\item{solution}{numeric; contains right answer for this numeric entry}

\item{score}{numeric, optional; the number of points for this gap; default 1}

\item{expected_length}{numeric, optional; is responsible to set a size of
text input field in content delivery engine}

\item{placeholder}{string, optional; is responsible to place some helpful
text in text input field in content delivery engine}

\item{tolerance}{numeric, optional; specifies the value for up and low
boundaries of tolerance rate for candidate answer}

\item{tolerance_type}{string, optional; specifies tolerance mode; possible
values:"exact", "absolute", "relative"}

\item{include_lower_bound}{boolean, optional; specifies whether or not the
lower bound is included in tolerance rate}

\item{include_upper_bound}{boolean, optional; specifies whether or not the
upper bound is included in tolerance rate}

\item{response_identifier}{string; an identifier for thy answer; by default
it is generated automatically}
}
\value{
string; map yaml
}
\description{
Create YAML string for NumericGap object
}
