% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gap_text}
\alias{gap_text}
\title{Create YAML string for TextGap object}
\usage{
gap_text(
  solution,
  alternatives = NULL,
  score = NULL,
  expected_length = NULL,
  placeholder = NULL,
  case_sensitive = NULL,
  response_identifier = NULL,
  tolerance = NULL
)
}
\arguments{
\item{solution}{string; contains right answer for this text gap entry}

\item{alternatives}{string vector, optional; contains a vector of values that
are also considered correct answers}

\item{score}{numeric, optional; the number of points for this gap; default 1}

\item{expected_length}{numeric, optional; is responsible to set a size of
text input field in content delivery engine}

\item{placeholder}{string, optional; is responsible to place some helpful
text in text input field in content delivery engine}

\item{case_sensitive}{logical, optional; determines whether the evaluation of
the correct answer is case sensitive}

\item{response_identifier}{string; an identifier for the answer; by default
it is generated automatically}

\item{tolerance}{numeric, optional; defines how many characters will be
taken into account to tolerate spelling mistake in evaluation of candidate
answer}
}
\value{
string; map yaml
}
\description{
Create YAML string for TextGap object
}
