% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{get_course_results}
\alias{get_course_results}
\title{Get course results by course id and node id}
\usage{
get_course_results(
  course_id,
  node_id,
  path = ".",
  api_user = NULL,
  api_password = NULL,
  endpoint = NULL
)
}
\arguments{
\item{course_id}{A length one character vector with course id.}

\item{node_id}{A length one character vector with node id (test).}

\item{path}{A length one character vector with path, where the zip should be
stored. Default is working directory.}

\item{api_user}{A character value of the username in the OPAL.}

\item{api_password}{A character value of the password in the OPAL.}

\item{endpoint}{A string of endpoint of LMS Opal; by default it is got from
environment variable \code{RQTI_API_ENDPOINT}. To set a global environment
variable, you need to call \code{Sys.setenv(RQTI_API_ENDPOINT='xxxxxxxxxxxxxxx')}
or you can put these command into .Renviron.}
}
\value{
A dataframe with the results of the course.
}
\description{
Get course results by course id and node id
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- get_course_results("89068111333293", "1617337826161777006")
\dontshow{\}) # examplesIf}
}
