% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{get_resource_url}
\alias{get_resource_url}
\title{Create a URL using the resource's display name in LMS OPAL}
\usage{
get_resource_url(
  display_name,
  endpoint = NULL,
  api_user = NULL,
  api_password = NULL
)
}
\arguments{
\item{display_name}{character; target display_name}

\item{endpoint}{endpoint of LMS Opal; default is from environment variable
\code{QTI_API_ENDPOINT}. To set a global environment variable, you need to call \verb{ }Sys.setenv(QTI_API_ENDPOINT='xxxxxxxxxxxxxxx')` or you can put these
commands into .Renviron.}

\item{api_user}{username on OPAL}

\item{api_password}{password on OPAL}
}
\description{
Create a URL using the resource's display name in LMS OPAL
}
