% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{get_resources}
\alias{get_resources}
\title{Get list of all user's resources on LMS OPAL}
\usage{
get_resources(
  endpoint = "https://bildungsportal.sachsen.de/opal/",
  api_user = NULL,
  api_password = NULL
)
}
\arguments{
\item{endpoint}{endpoint of LMS Opal; default is
"https://bildungsportal.sachsen.de/opal/"}

\item{api_user}{username on OPAL}

\item{api_password}{password on OPAL}
}
\description{
Get list of all user's resources on LMS OPAL
}
