% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentSection.R,
%   R/character.R
\docType{methods}
\name{getAssessmentItems}
\alias{getAssessmentItems}
\alias{getAssessmentItems,AssessmentItem-method}
\alias{getAssessmentItems,AssessmentItem}
\alias{getAssessmentItems,AssessmentSection-method}
\alias{getAssessmentItems,AssessmentSection}
\alias{getAssessmentItems,character-method}
\alias{getAssessmentItems,character}
\title{Get list of AssessmentItems for AssessmentSection}
\usage{
getAssessmentItems(object)

\S4method{getAssessmentItems}{AssessmentItem}(object)

\S4method{getAssessmentItems}{AssessmentSection}(object)

\S4method{getAssessmentItems}{character}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{AssessmentSection},
\link{AssessmentItem})}
}
\description{
Generic function for
}
