% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentSection.R,
%   R/character.R
\docType{methods}
\name{getContributors-methods}
\alias{getContributors-methods}
\alias{getContributors}
\alias{getContributors,AssessmentItem-method}
\alias{getContributors,AssessmentItem}
\alias{getContributors,AssessmentSection-method}
\alias{getContributors,AssessmentSection}
\alias{getContributors,character-method}
\alias{getContributors,character}
\title{Get list of contributors values}
\usage{
getContributors(object)

\S4method{getContributors}{AssessmentItem}(object)

\S4method{getContributors}{AssessmentSection}(object)

\S4method{getContributors}{character}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Ordering}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Get list of contributors values
}
