% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R, R/Opal.R
\name{getCourseAssessment}
\alias{getCourseAssessment}
\alias{getCourseAssessment,missing-method}
\alias{getCourseAssessment,Opal-method}
\title{Get assessment scores for a course element}
\usage{
getCourseAssessment(object, course_id, node_id, user_id = NULL, ...)

\S4method{getCourseAssessment}{missing}(object, course_id, node_id, user_id = NULL, ...)

\S4method{getCourseAssessment}{Opal}(object, course_id, node_id, user_id = NULL)
}
\arguments{
\item{object}{An S4 object of class \link{LMS} that represents a connection to the LMS.}

\item{course_id}{A length one character vector with course id.}

\item{node_id}{A length one character vector with course element id.}

\item{user_id}{A length one character vector with a user login name or email
address. If \code{NULL}, assessments are returned for all users.}

\item{...}{Additional arguments to be passed to the method, if applicable.}
}
\value{
A data frame with assessment scores.
}
\description{
This method retrieves current assessment data for a course element on a
Learning Management System (LMS), including score, maximum score, passed
status, and attempts. If no LMS connection object is provided, it attempts to
guess the connection using default settings (e.g., environment variables). If
the connection cannot be established, an error is thrown.
}
\examples{
\dontshow{if (interactive()) \{}
assessment <- getCourseAssessment("89068111333293", "1617337826161777006")
\dontshow{\}}
}
