% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R
\docType{methods}
\name{getFiles-methods}
\alias{getFiles-methods}
\alias{getFiles}
\alias{getFiles,AssessmentItem-method}
\alias{getFiles,AssessmentItem}
\title{Get file paths for attachment of test}
\usage{
getFiles(object)

\S4method{getFiles}{AssessmentItem}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Get file paths for attachment of test
}
