% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentSection.R,
%   R/AssessmentTest.R, R/Gap.R
\docType{methods}
\name{getIdentifier-methods}
\alias{getIdentifier-methods}
\alias{getIdentifier}
\alias{getIdentifier,AssessmentItem-method}
\alias{getIdentifier,AssessmentItem}
\alias{getIdentifier,AssessmentSection-method}
\alias{getIdentifier,AssessmentSection}
\alias{getIdentifier,character-method}
\alias{getIdentifier,character}
\alias{getIdentifier,Gap-method}
\alias{getIdentifier,Gap}
\title{Get identifier}
\usage{
getIdentifier(object)

\S4method{getIdentifier}{AssessmentItem}(object)

\S4method{getIdentifier}{AssessmentSection}(object)

\S4method{getIdentifier}{character}(object)

\S4method{getIdentifier}{Gap}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Get identifier
}
