% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentItem.R, R/AssessmentSection.R,
%   R/MultipleChoice.R, R/character.R
\docType{methods}
\name{getPoints-methods}
\alias{getPoints-methods}
\alias{getPoints}
\alias{getPoints,AssessmentItem-method}
\alias{getPoints,AssessmentItem}
\alias{getPoints,AssessmentSection-method}
\alias{getPoints,AssessmentSection}
\alias{getPoints,MultipleChoice-method}
\alias{getPoints,MultipleChoice}
\alias{getPoints,character-method}
\alias{getPoints,character}
\title{Get points from AssessmentItem object}
\usage{
getPoints(object)

\S4method{getPoints}{AssessmentItem}(object)

\S4method{getPoints}{AssessmentSection}(object)

\S4method{getPoints}{MultipleChoice}(object)

\S4method{getPoints}{character}(object)
}
\arguments{
\item{object}{an instance of the S4 object (\link{SingleChoice}, \link{MultipleChoice},
\link{Essay}, \link{Entry}, \link{Order}, \link{OneInRowTable}, \link{OneInColTable},
\link{MultipleChoiceTable}, \link{DirectedPair}, \link{TextGap}, \link{NumericGap},
\link{InlineChoice})}
}
\description{
Get points from AssessmentItem object
}
