% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_functions.R
\name{knit_qti_html}
\alias{knit_qti_html}
\title{Render an RMD file as qti xml with QTIJS}
\usage{
knit_qti_html(input, ...)
}
\arguments{
\item{input}{(the path to the input Rmd document)}

\item{...}{required for passing arguments when knitting}
}
\description{
Generates the qti xml file through the functions of the package exams. The
xml is copied into the QTIJS folder which transforms the xml into HTML.
Finally, the HTML is displayed and the user can have a preview of the
exercise or exam.
}
\details{
Requires a running QTIJS server, which can be started with start_server().
}
\examples{
# Customize knit function in the Rmd file using the following YAML setting
# after the word knit:
\dontrun{qti::knit}

}
