% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchTable.R
\docType{class}
\name{MatchTaable-classs}
\alias{MatchTaable-classs}
\alias{MatchTable}
\title{Class "MatchTable"}
\description{
Abstract class \code{MatchTable} is never to be generated, only derived classes
OneInRowTable, OneInColTable, MultipleChoiceTable and DirectedPair are
meaningful.
}
\section{Slots}{

\describe{
\item{\code{rows}}{string vector; sets answer options as row names in table or the
firsts elements in couples in directed pairs}

\item{\code{rows_identifiers}}{string vector; sets identifiers for answer options
defined in rows of table or identifiers of the firsts elements in couples
in directed pairs}

\item{\code{cols}}{string vector; sets answer options as cols headers in table or the
seconds elements in couples in directed pairs}

\item{\code{cols_identifiers}}{string vector; sets identifiers for answer options
defined in cols of table or identifiers of the seconds elements in couples
in directed pairs}

\item{\code{answers_identifiers}}{string vector; sets couples of identifiers that
combine the right answers}

\item{\code{answers_scores}}{numeric vector, optional; each number in this vector
determines the number of points that will be awarded to a candidate if he
selects the corresponding answer; if nothing assign as answers_scores, the
individual values for right answers calculates from task points and number
of right options}

\item{\code{shuffle}}{boolean, optional; is responsible to randomize the order in
which the choices are initially presented to the candidate; default \code{TRUE}}

\item{\code{shuffle_rows}}{boolean, optional; shuffle possible answers in rows,
default \code{TRUE}}

\item{\code{shuffle_cols}}{boolean, optional; shuffle possible answers in columns;
defalut \code{TRUE}}
}}

\section{Warning}{
 This class is not useful in itself, but some classes derive
from it.
}

