% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalFeedback.R
\docType{class}
\name{ModalFeedback-class}
\alias{ModalFeedback-class}
\alias{ModalFeedback}
\title{Class ModalFeedback}
\description{
Abstract class \code{ModalFeedback} is responsible for delivering feedback
messages to the candidate, regardless of whether the answer was correct or
incorrect
}
\section{Slots}{

\describe{
\item{\code{outcome_identifier}}{string; identifier if outcome declaration variable
that relates to feedback}

\item{\code{show}}{boolean, optional; is responsible to show (\code{TRUE} value) or hide
(\code{FALSE}); default \code{TRUE}}

\item{\code{identifier}}{string; identifier of model feedback item; default 'modal
feedback'}

\item{\code{title}}{string, optional; title of model feedback item}

\item{\code{content}}{list; list of strings with formatting to form the text of the
question}
}}

