% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleChoice.R
\docType{class}
\name{MultipleChoice-class}
\alias{MultipleChoice-class}
\alias{MultipleChoice}
\title{Class "MultipleChoice"}
\description{
Abstract class \code{MuiltipleChoice} is responsible for creating multiple choice
assessment task according to QTI 2.1.
}
\section{Slots}{

\describe{
\item{\code{content}}{list; list of strings with formatting to form the text of the
question}

\item{\code{identifier}}{string; identifier of assessment task}

\item{\code{title}}{string; a title of xml file}

\item{\code{prompt}}{string, optional; simple question text, consisting of one
paragraph, can supplement or replace content}

\item{\code{qti_version}}{string; QTI information model version; default 'v2p1}

\item{\code{feedback}}{list of \linkS4class{ModalFeedback}, \link{CorrectFeedback}, or
\linkS4class{WrongFeedback}; feedback messages for candidate}

\item{\code{files}}{string, optional; the names of the attachment files that should
be available in the test}

\item{\code{calculator}}{string, optional; determines to show to candidate calculator
during the test;
possible values:
\itemize{
\item "simple-calculator"
\item "scientific-calculator"
\item "" - by default
}}

\item{\code{choices}}{string vector; defines a set of answers options in question}

\item{\code{shuffle}}{boolean, optional; is responsible to randomize the order in
which the choices are initially presented to the candidate; default \code{TRUE}}

\item{\code{choice_identifiers}}{string vector, optional; a set of identifiers for
answers; by default identifiers are generated automatically}

\item{\code{orientation}}{string, optional; is responsible to place answers in
vertical or horizontal mode; possible values:
\itemize{
\item "vertical" - default
\item "horizontal"
}}

\item{\code{points}}{numeric vector, required; each number in this vector determines
the number of points that will be awarded to a candidate if he selects the
corresponding answer; the order of the scores must follow the order of the
\code{choices}; it is possible to assign negative values to incorrect answers;
all answers with a positive score are considered correct}
}}

\note{
use \code{\link[=create_qti_task]{create_qti_task()}} to create an xml file of a task specification
}
\examples{
mc <- new("MultipleChoice",
          content = list("<p>Pick up the right options</p>"),
          choices = c("option 1", "option 2", "option 3", "option 4"),
          title = "single_choice_task",
          shuffle = FALSE,
          points = c(0.5, 0,5, 0, 0),
          identifier = "mc_example")
}
