% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleChoice.R
\docType{class}
\name{MultipleChoice-class}
\alias{MultipleChoice-class}
\alias{MultipleChoice}
\title{Class "MultipleChoice"}
\description{
Abstract class \code{MuiltipleChoice} is responsible for creating multiple choice
assessment task according to QTI 2.1.
}
\section{Slots}{

\describe{
\item{\code{content}}{list; a list of content to form text of question, includes text
(characters) or instances of Gap-objects: \linkS4class{TextGap}, \link{NumericGap},
InlineChoice}

\item{\code{identifier}}{string; identifier of assessment task}

\item{\code{title}}{string; a title of xml file}

\item{\code{prompt}}{string, optional; simple question text, consisting of one
paragraph, can supplement or replace content}

\item{\code{qti_version}}{string; QTI information model version; default 'v2p1'}

\item{\code{choices}}{string vector; define a set of answer options in question}

\item{\code{shuffle}}{boolean, optional; is responsible to randomize the order in
which the choices are initially presented to the candidate; default \code{TRUE}}

\item{\code{choice_identifiers}}{string vector, optional; a set of identifiers for
answers; by default identifiers are generated automatically}

\item{\code{orientation}}{string, optional; is responsible to place answers in
vertical or horizontal mode; possible values:
\itemize{
\item default \code{""}
\item "vertical"
\item "horizontal"
}}

\item{\code{points}}{numeric vector, required; each number in this vector determines
the number of points that will be awarded to a candidate if he selects the
corresponding answer; the order of the scores must follow the order of the
\code{choices}; it is possible to assign negative values to incorrect answers;
all answers with a positive score are considered correct}

\item{\code{mapping}}{do not use; value initialize automatically; named numeric
vector of points, where names are choice identifiers}

\item{\code{lower_bound}}{numeric, optional; min number of options, that candidate
has to choose; default \code{0}}

\item{\code{upper_bound}}{do not use;value initialize automatically; define max
number of right options in the answer set}

\item{\code{default_value}}{do not use; the default value from the target set to be
used when no explicit mapping for a source value is given; default \code{0}}

\item{\code{maxscore}}{do not use; value initialize automatically as sum of positive
points}
}}

\examples{
mc <- new("MultipleChoice",
          content = list("<p>Pick up the right options</p>"),
          choices = c("option 1", "option 2", "option 3", "option 4"),
          title = "single_choice_task",
          shuffle = FALSE,
          points = c(0.5, 0,5, 0, 0),
          identifier = "mc_example")
}
