% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericGap.R
\docType{class}
\name{NumericGap-class}
\alias{NumericGap-class}
\alias{NumericGap}
\title{Class NumericGap}
\description{
Abstract class \code{NumericGap} is responsible for creating instances of input
fields with numeric type of answer in question Entry type assessment task
according to QTI 2.1.
}
\section{Slots}{

\describe{
\item{\code{response_identifier}}{string; an identifier for thy answer; by default it
is generated automatically}

\item{\code{points}}{numeric, optional; the number of points for this gap;
default \code{1}}

\item{\code{placeholder}}{string, optional; is responsible to place some helpful text
in text input field in content delivery engine}

\item{\code{expected_length}}{numeric, optional; is responsible to set a size of text
input field in content delivery engine}

\item{\code{solution}}{numeric; contains right answer for this numeric entry}

\item{\code{tolerance}}{numeric, optional; specifies the value for up and low
boundaries of tolerance rate for candidate answer}

\item{\code{tolerance_type}}{string, optional; specifies tolerance mode; possible
values:
\itemize{
\item "exact" by default
\item "absolute"
\item "relative"
}}

\item{\code{include_lower_bound}}{boolean, optional; specifies whether or not the
lower bound is included in tolerance rate}

\item{\code{include_upper_bound}}{boolean, optional; specifies whether or not the
upper bound is included in tolerance rate}
}}

\examples{
ng <- new("NumericGap",
          response_identifier = "gap_1",
          solution = 5,
          placeholder = "use this format xx.xxx" )
}
