% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opal.R
\docType{class}
\name{Opal-class}
\alias{Opal-class}
\alias{Opal}
\title{Class Opal}
\description{
The \code{Opal} class represents a specific implementation of a Learning Management
System (LMS) that extends the abstract \code{LMS} class. This class is designed to
facilitate interactions with the Opal LMS API.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string representing the name/identifier of the LMS.
Defaults to \code{"Opal"}.}

\item{\code{api_user}}{A character string specifying the API username for authentication.}

\item{\code{endpoint}}{A character string containing the API endpoint of the Opal LMS.
This can be set using the environment variable \code{RQTI_API_ENDPOINT} with
\code{Sys.setenv(RQTI_API_ENDPOINT='xxxxxxxxxxxxxxx')} or placed in the \code{.Renviron} file.}
}}

\seealso{
\link{LMS-class} for the parent class.
}
