% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opal.R
\name{opal}
\alias{opal}
\title{Create an Opal LMS Connection Object}
\usage{
opal(api_user = NA_character_, endpoint = NA_character_)
}
\arguments{
\item{api_user}{A character string specifying the API username.}

\item{endpoint}{A character string specifying the API endpoint for the LMS.}
}
\value{
An object of class \code{Opal}, inheriting from \code{LMS},
which can be used to interact with the Opal LMS API.
}
\description{
This helper function initializes an \code{Opal} object, a subclass of \code{LMS},
representing a connection to the Opal Learning Management System (LMS).
}
