% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order.R
\docType{class}
\name{Order-class}
\alias{Order-class}
\alias{Order}
\title{Class "Order"}
\description{
Abstract class \code{Order} is responsible for creating assessment task according
to QTI 2.1., where candidate has to place answers in a specific order
}
\section{Slots}{

\describe{
\item{\code{content}}{list; list of strings with formatting to form the text of the
question}

\item{\code{identifier}}{string; identifier of assessment task}

\item{\code{title}}{string; a title of xml file}

\item{\code{prompt}}{string, optional; simple question text, consisting of one
paragraph, can supplement or replace content}

\item{\code{qti_version}}{string; QTI information model version; default 'v2p1}

\item{\code{feedback}}{list of \linkS4class{ModalFeedback}, \link{CorrectFeedback}, or
\linkS4class{WrongFeedback}; feedback messages for candidate}

\item{\code{files}}{string, optional; the names of the attachment files that should
be available in the test}

\item{\code{calculator}}{string, optional; determines to show to candidate calculator
during the test;
possible values:
\itemize{
\item "simple-calculator"
\item "scientific-calculator"
\item "" - by default
}}

\item{\code{choices}}{string vector; contains the answers; the order of answers in
the vector is a right response for the task}

\item{\code{choices_identifiers}}{string vector, optional; a set of identifiers for
answers; by default identifiers are generated automatically}

\item{\code{shuffle}}{boolean, optional; is responsible to randomize the order in
which the choices are initially presented to the candidate; default \code{TRUE}}

\item{\code{points_per_answer}}{boolean, optional; in case of \code{TRUE} each selected
answer will be scored individually, in case of \code{FALSE} only answer wich is
fully correct will be scored with maximum score; \code{TRUE} by default}

\item{\code{points}}{numeric, optional; the number of points for the entire task;
default 1}
}}

\note{
use \code{\link[=create_qti_task]{create_qti_task()}} to create an xml file of a task specification
}
\examples{
ord <- new("Order", content = list("<p>put in a right order</p>"),
           choices = c("first", "second", "third"),
           title = "order",
           identifier = "order_example")
}
