% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QtiMetadata.R
\docType{class}
\name{QtiMetadata-class}
\alias{QtiMetadata-class}
\alias{QtiMetadata}
\title{Class QtiMetadata}
\description{
This class stores metadata information such as contributors, description,
rights, version and format for QTI-compliant tasks or tests.
}
\section{Slots}{

\describe{
\item{\code{contributor}}{A list of objects \link{QtiContributor}-type that holds metadata
information about the authors.}

\item{\code{description}}{A character string providing a textual description of the
content of this learning object.}

\item{\code{rights}}{A character string describing the intellectual property rights
and conditions of use for this learning object. By default it takes value
from environment variable 'QTI_RIGHTS'.}

\item{\code{version}}{A character string representing the edition/version of this
learning object.}

\item{\code{format}}{A character string representing the QTI (Question and Test
Interoperability) information model version. Default is 'IMS QTI 2.1'.}
}}

