% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_functions.R
\name{render_opal}
\alias{render_opal}
\title{Render Rmd directly in Opal via API}
\usage{
render_opal(input, ...)
}
\arguments{
\item{input}{(the path to the input Rmd document)}

\item{...}{required for passing arguments when knitting}
}
\description{
Render Rmd directly in Opal via API
}
