% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_functions.R
\name{render_xml}
\alias{render_xml}
\title{Render a single xml file with QTIJS}
\usage{
render_xml(input)
}
\arguments{
\item{input}{input file}
}
\value{
nothing, has side effects
}
\description{
Uses QTIJS to render a single xml file in the RStudio viewer pane with a
local server.
}
