% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_functions.R
\name{render_zip}
\alias{render_zip}
\title{Render a zipped qti archive with QTIJS}
\usage{
render_zip(input)
}
\arguments{
\item{input}{input file}
}
\value{
nothing, has side effects
}
\description{
Uses QTIJS to render a zipped qti archive in the RStudio viewer pane with a
local server.
}
