% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_builder.R
\name{rmd2zip}
\alias{rmd2zip}
\title{Create test zip file with one task xml file from Rmd (md) description}
\usage{
rmd2zip(file, path = getwd(), verification = FALSE)
}
\arguments{
\item{file}{A string of path to file with markdown description of question.}

\item{path}{A string, optional; a folder to store xml file. Default is
working directory.}

\item{verification}{A boolean value, optional; enable validation of the xml
file. Default is \code{FALSE}.}
}
\value{
The path string to the zip file.
}
\description{
Create zip file with test, that contains one xml question specification
generated from Rmd (md) description according to qti 2.1 information model
}
\examples{
\dontrun{
rmd2zip("task.Rmd", "target_folder", TRUE)
}
}
