% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqti-package.R
\docType{package}
\name{rqti-package}
\alias{rqti}
\alias{rqti-package}
\title{rqti: Create Tests According to QTI 2.1 Standard}
\description{
Create tests and tasks compliant with the Question & Test Interoperability (QTI) information model version 2.1. Input sources are Rmd/md description files or S4-class objects. Output formats include standalone zip or xml files. Supports the generation of basic task types (single and multiple choice, order, pair association, matching tables, filling gaps and essay) and provides a comprehensive set of attributes for customizing tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shevandrin/rqti}
  \item \url{https://shevandrin.github.io/rqti/}
  \item Report bugs at \url{https://github.com/shevandrin/rqti/issues}
}

}
\author{
\strong{Maintainer}: Andrey Shevandrin \email{shevandrin@gmail.com} (\href{https://orcid.org/0000-0003-0807-2546}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Johannes Titz \email{johannes@titz.science} (\href{https://orcid.org/0000-0002-1102-5719}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Petr Bondarenko \email{bondarenko.volsu@gmail.com} (\href{https://orcid.org/0000-0002-5154-9664}{ORCID}) [contributor]
  \item Ivonne Ojeda \email{yulieth9109@gmail.com} [contributor]
  \item Brian Mottershead (Author of QTIJS library) [copyright holder]
  \item Stiftung für Innovation in der Hochschullehre [funder]
}

}
\keyword{internal}
