% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleChoice.R
\docType{class}
\name{SingleChoice-class}
\alias{SingleChoice-class}
\alias{SingleChoice}
\title{Class "SingleChoice"}
\description{
Abstract class \code{SingleChoice} is responsible for creating single choice
assessment task according to QTI 2.1.
}
\section{Slots}{

\describe{
\item{\code{content}}{list; a list of content to form text of question, includes text
(characters) or instances of Gap-objects: \linkS4class{TextGap}, \link{NumericGap},
InlineChoice}

\item{\code{identifier}}{string; identifier of assessment task}

\item{\code{title}}{string; a title of xml file}

\item{\code{prompt}}{string, optional; simple question text, consisting of one
paragraph, can supplement or replace content}

\item{\code{qti_version}}{string; QTI information model version; default 'v2p1'}

\item{\code{choices}}{string vector; define a set of answer options in question}

\item{\code{shuffle}}{boolean, optional; is responsible to randomize the order in
which the choices are initially presented to the candidate; default \code{TRUE}}

\item{\code{choice_identifiers}}{string vector, optional; a set of identifiers for
answers; by default identifiers are generated automatically}

\item{\code{orientation}}{string, optional; is responsible to place answers in
vertical or horizontal mode; possible values:
\itemize{
\item default \code{""}
\item "vertical"
\item "horizontal"
}}

\item{\code{solution}}{numeric, optional; the number of right answer in choices slot,
by default the first item in choices slot is considered as right answer;
default \code{1}}

\item{\code{points}}{numeric, optional; the number of points for the entire task;
default 1}
}}

\examples{
sc <- new("SingleChoice",
          content = list("<p>Pick up the right option</p>"),
          choices = c("option 1", "option 2", "option 3", "option 4"),
          title = "single_choice_task",
          shuffle = FALSE,
          points = 2,
          identifier = "sc_example")
}
