% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleChoice.R
\name{singleChoice}
\alias{singleChoice}
\title{Create object of "\linkS4class{SingleChoice}" class}
\usage{
singleChoice(
  identifier,
  title,
  choices,
  choice_identifiers = NA_character_,
  solution = NA_integer_,
  content = list(),
  points = NA_integer_,
  shuffle = logical(0L),
  prompt = NA_character_,
  orientation = NA_character_
)
}
\arguments{
\item{identifier}{a character string; some identifier of question}

\item{title}{a character string; some title of question}

\item{choices}{a character vector; set of possible answers; by default the
first one is considered as a right response}

\item{choice_identifiers}{a character vector, optional; set of identifiers of
answer options; can be assigned automatically}

\item{solution}{a number, optional; the index of the right answer in vector
of choices}

\item{content}{a list of character items; this character items should be
formatted text that combines question/task description}

\item{points}{a number, optional; the number of points for the right answer}

\item{shuffle}{logical, optional; option to mix answer options up, default
'TRUE'}

\item{prompt}{a character string, optional; text question in one paragraph,
alternative to 'text' parameter}

\item{orientation}{a character string with two possible values of "vertical"
(default) or "horizontal" placement of answer options}
}
\value{
an object of S4 class "\linkS4class{SingleChoice}"
}
\description{
Create object of "\linkS4class{SingleChoice}" class
}
\examples{
sc <- singleChoice(title = "single choice task",
                   content = list("<p>Pick up the right option</p>"),
                   choices= c("option 1", "option 2", "option 3"),
                   identifier = "sc_example")

}
