% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_functions.R
\name{start_server}
\alias{start_server}
\title{Start QTIJS on a local server}
\usage{
start_server()
}
\value{
The URL string for QTIJS server.
}
\description{
This function starts an http server with the QTIJS renderer. The renderer performs the conversion of qti.xml into HTML.
}
\details{
The server has to be started manually by the user, otherwise the Knit Button will not work. The Button starts a new session and invoking a server there does not make much sense.
}
\examples{
\dontrun{
# Initiated server in qtiViewer folder
start_server()
# Initiated server in a specific folder provided by the user. This folder
# contains the QTI renderer
start_server("/pathToTheQtiRenderer/")
}
}
