% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_builder.R
\name{test}
\alias{test}
\title{Create test}
\usage{
test(
  content,
  identifier = NULL,
  title = NULL,
  navigation_mode = "nonlinear",
  submission_mode = "individual",
  time_limits = NULL,
  max_attempts = NULL,
  allow_comment = TRUE,
  rebuild_variables = TRUE
)
}
\arguments{
\item{content}{list contains \link{AssessmentSection} objects}

\item{identifier}{string; identifier if the test file}

\item{title}{string, optional; file title}

\item{navigation_mode}{string, optional; determines the general paths that the
candidate may have during exam; two mode options are possible: \code{linear}
"linear" - candidate is not allowed to return to the previous questions;
\code{nonlinear}- candidate if free to navigate, used by default}

\item{submission_mode}{string, optional; determines when the candidate's responses
are submitted for response processing; two mode options are possible:
\code{individual} - submit candidates' responses on an item-by-idem basis, used
by default; \code{simultaneous - candidates} - responses are submitted all
together by the end of the test}

\item{time_limits}{numeric, optional; controls the amount of time a candidate
is given for the test}

\item{max_attempts}{numeric, optional; enables the maximum number of
attempts, that candidate is allowed to pass}

\item{allow_comment}{boolean, optional; enables to allow candidate to leave
comments in each question, \code{TRUE} by default}

\item{rebuild_variables}{boolean, optional; enables to recalculate variables and
reshuffle the order of choices for each item-attempt}
}
\description{
Create test
}
