% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_builder.R
\name{test}
\alias{test}
\title{Create test}
\usage{
test(
  content,
  identifier = "test_identifier",
  title = "Test Title",
  time_limit = 90,
  max_attempts = 1,
  academic_grading = FALSE,
  grade_label = c(en = "Grade", de = "Note"),
  table_label = c(en = "Grade", de = "Note"),
  navigation_mode = "nonlinear",
  submission_mode = "individual",
  allow_comment = TRUE,
  rebuild_variables = TRUE
)
}
\arguments{
\item{content}{A list containing \link{AssessmentSection} objects.}

\item{identifier}{A string indicating the identifier of the test file. Default is 'test_identifier'.}

\item{title}{A string, optional, representing the file title. Default is
'Test Title'.}

\item{time_limit}{A numeric value, optional, controlling the time given to a
candidate for the test in minutes. Default is 90 minutes.}

\item{max_attempts}{A numeric value, optional, indicating the maximum number
of attempts allowed for the candidate. Default is 1.}

\item{academic_grading}{A boolean, optional; enables to show to candidate at the
end of the testing a grade according to 5-point academic grade system as a
feedback; \code{FALSE} by default.}

\item{grade_label}{A string, optional; a short message that shows with a
grade in the final feedback; for multilingual use, it can be a named vector
with two-letter ISO language codes as names (e.g., c(en="Grade",
de="Note")); during test creation, it takes the value for the language of
the operating system; c(en="Grade", de="Note") is default}

\item{navigation_mode}{A string, optional, determining the general paths that
the candidate may have during the exam. Two mode options are possible:
\itemize{
\item 'linear': Candidate is not allowed to return to previous questions.
\item 'nonlinear': Candidate is free to navigate; used by default.
}}

\item{submission_mode}{A string, optional, determining when the candidate's
responses are submitted for response processing. One of two mode options is
possible:
\itemize{
\item 'individual': Submit candidates' responses on an item-by-item basis; used by default.
\item 'simultaneous': Candidates' responses are submitted all together by the end of the test.
}}

\item{allow_comment}{A boolean, optional, enabling the candidate to leave
comments in each question. \code{TRUE} by default.}

\item{rebuild_variables}{A boolean, optional, enabling the recalculation of
variables and reshuffling the order of choices for each item-attempt. \code{TRUE}
by default.}
}
\value{
An \link{AssessmentTest} object
}
\description{
Create test
}
\section{Slots}{

\describe{
\item{\code{table_label}}{A string, optional; a concise message to display as the
column title of the grading table in the final feedback; for multilingual use,
it can be a named vector with two-letter ISO language codes as names
(e.g., c(en="Grade", de="Note")); during test creation, it takes the value
for the language of the operating system; c(en="Grade", de="Note")is default.}
}}

