% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_builder.R
\name{test4opal}
\alias{test4opal}
\title{Create test for LMS Opal}
\usage{
test4opal(
  content,
  identifier = "test_identifier",
  title = "Test Title",
  time_limit = 90,
  max_attempts = 1,
  files = NULL,
  calculator = "scientific-calculator",
  navigation_mode = "nonlinear",
  submission_mode = "individual",
  allow_comment = TRUE,
  rebuild_variables = TRUE,
  show_test_time = TRUE,
  mark_items = TRUE,
  keep_responses = FALSE
)
}
\arguments{
\item{content}{A list containing \link{AssessmentSection} objects.}

\item{identifier}{A string indicating the identifier of the test file.
Default is 'test_identifier'.}

\item{title}{A string, optional, representing the file title. Default is
'Test Title'.}

\item{time_limit}{A numeric value, optional, controlling the time given to a
candidate for the test in minutes. Default is 90 minutes.}

\item{max_attempts}{A numeric value, optional, indicating the maximum number
of attempts allowed for the candidate. Default is 1.}

\item{files}{A character vector, optional; paths to files that will be
accessible to the candidate during the test/exam.}

\item{calculator}{A character, optional; determines whether to show a calculator to the candidate. Possible values:
\itemize{
\item 'simple-calculator'
\item 'scientific-calculator' (assigned by default)
}}

\item{navigation_mode}{A string, optional, determining the general paths that
the candidate may have during the exam. Two mode options are possible:
\itemize{
\item 'linear': Candidate is not allowed to return to previous questions.
\item 'nonlinear': Candidate is free to navigate; used by default.
}}

\item{submission_mode}{A string, optional, determining when the candidate's
responses are submitted for response processing. One of two mode options is
possible:
\itemize{
\item 'individual': Submit candidates' responses on an item-by-item basis; used by default.
\item 'simultaneous': Candidates' responses are submitted all together by the end of the test.
}}

\item{allow_comment}{A boolean, optional, enabling the candidate to leave
comments in each question. \code{TRUE} by default.}

\item{rebuild_variables}{A boolean, optional, enabling the recalculation of
variables and reshuffling the order of choices for each item-attempt. \code{TRUE}
by default.}

\item{show_test_time}{A boolean, optional, determining whether to show
candidate elapsed processing time without a time limit. Default is \code{TRUE}.}

\item{mark_items}{A boolean, optional, determining whether to allow candidate
marking of questions. Default is \code{TRUE}.}

\item{keep_responses}{A boolean, optional, determining whether to save the
candidate's answers from the previous attempt. Default is \code{FALSE}.}
}
\value{
An \link{AssessmentTestOpal} object
}
\description{
Create test for LMS Opal
}
