% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextGapOpal.R
\docType{class}
\name{TextGapOpal-class}
\alias{TextGapOpal-class}
\alias{TextGapOpal}
\title{Class TextGapOpal}
\description{
Abstract class \code{TextGapOpal} is responsible for creating instances of input
fields with text type of answer in question Entry type assessment task
according to QTI 2.1. for LMS Opal
}
\section{Slots}{

\describe{
\item{\code{response_identifier}}{string; an identifier for thy answer; by default it
is generated automatically}

\item{\code{score}}{numeric, optional; the number of points for this gap;
default \code{1}}

\item{\code{placeholder}}{string, optional; is responsible to place some helpful text
in text input field in content delivery engine}

\item{\code{expected_length}}{numeric, optional; is responsible to set a size of text
input field in content delivery engine}

\item{\code{solution}}{string; contains a vector of values that are
considered as correct answers}

\item{\code{case_sensitive}}{logical; determines whether the evaluation of the
correct answer is case sensitive; default \code{TRUE}}

\item{\code{tolerance}}{numeric; defines how many characters will be taken into
account to tolerate spelling mistake in evaluation of candidate answer}
}}

\examples{
tgo <- new("TextGapOpal",
          solution = c("answer","answerr", "aanswer"),
          placeholder = "do not put special characters",
          tolerance = 1)
}
