% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{upload2opal}
\alias{upload2opal}
\title{Upload resource on OPAL}
\usage{
upload2opal(
  file,
  display_name = NULL,
  access = 4,
  endpoint = paste0("https://bildungsportal.sachsen.de/", "opal/"),
  in_browser = TRUE
)
}
\arguments{
\item{file}{required; a length one character vector}

\item{display_name}{optional; a length one character vector to entitle file in
OPAL; file name without extension by default}

\item{access}{optional; is responsible for publication status, where 1 - only
those responsible for this learning resource; 2 - responsible and other
authors; 3 - all registered users; 4 - default value, registered users and
guests}

\item{endpoint}{endpoint}

\item{in_browser}{logical, optional; the parameter that controls whether to
open a URL in default browser; TRUE by default}
}
\value{
status code
}
\description{
Function \code{upload2opal()} takes full prepared zip archive of QTI-test or
QTI-task and uploads it to the OPAL. before calling \code{upload2opal()}
authentication procedure has to be performed. See \link{auth_opal}
}
