% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_opal.R
\name{upload2opal}
\alias{upload2opal}
\title{Upload resource on OPAL}
\usage{
upload2opal(
  test,
  display_name = NULL,
  access = 4,
  overwrite = TRUE,
  endpoint = NULL,
  open_in_browser = TRUE,
  api_user = NULL,
  api_password = NULL
)
}
\arguments{
\item{test}{required; a length one character vector or \link{AssessmentTest} or
\link{AssessmentTestOpal} objects}

\item{display_name}{optional; a length one character vector to entitle file in
OPAL; file name without extension by default}

\item{access}{optional; is responsible for publication status, where 1 - only
those responsible for this learning resource; 2 - responsible and other
authors; 3 - all registered users; 4 - default value, registered users and
guests}

\item{overwrite}{logical; if only one file with the specified display name is
found, it will be overwritten}

\item{endpoint}{endpoint}

\item{open_in_browser}{logical, optional; the parameter that controls whether
to open a URL in default browser; TRUE by default}

\item{api_user}{username on OPAL}

\item{api_password}{password on OPAL}
}
\value{
list with key and url
}
\description{
Function \code{upload2opal()} takes full prepared zip archive of QTI-test or
QTI-task and uploads it to the OPAL. before calling \code{upload2opal()}
authentication procedure has to be performed. See \link{auth_opal}
}
\section{Authentication}{
 To use OPAL API, you need to provide your
OPAL-username and password. This function can get \code{api_user} from
environment variables. To set a global environment variable, you need to
call \code{Sys.setenv(QTI_API_USER ='xxxxxxxxxxxxxxx')} or you can put these
commands into .Renviron. This package uses system credential store 'keyring'
to store user's password.
}

